### 2025-02-20 bbea313
kwave 3D source offset z vec bug fix
