### 2025-02-20 766706f
joss paper references cleanup
