+++ joss_paper/references.bib
@@ -1,0 +1,344 @@
+@article{Harvey1929,
+  title = {The Effect of High Frequency Sound Waves on Heart Muscle and Other Irritable Tissues},
+  author = {Harvey, E. Newton},
+  year = {1929},
+  month = dec,
+  journal = {American Journal of Physiology-Legacy Content},
+  volume = {91},
+  number = {1},
+  pages = {284--290},
+  publisher = {American Physiological Society},
+  doi = {10.1152/ajplegacy.1929.91.1.284},
+  file = {/Users/tomaubier/Library/Mobile Documents/iCloud~md~obsidian/Documents/UsNeuStim/Files/Harvey1929.pdf}
+}
+
+@article{Dyke2017,
+  title = {Comparing {{GABA-dependent}} Physiological Measures of Inhibition with Proton Magnetic Resonance Spectroscopy Measurement of {{GABA}} Using Ultra-High-Field {{MRI}}},
+  author = {Dyke, Katherine and P{\'e}p{\'e}s, Sophia E. and Chen, Chen and Kim, Soyoung and Sigurdsson, Hilmar P. and Draper, Amelia and Husain, Masud and Nachev, Parashkev and Gowland, Penelope A. and Morris, Peter G. and Jackson, Stephen R.},
+  year = {2017},
+  month = may,
+  journal = {NeuroImage},
+  volume = {152},
+  pages = {360--370},
+  issn = {10538119},
+  doi = {10.1016/j.neuroimage.2017.03.011},
+  urldate = {2025-01-27},
+  abstract = {Imbalances in glutamatergic (excitatory) and GABA (inhibitory) signalling within key brain networks are thought to underlie many brain and mental health disorders, and for this reason there is considerable interest in investigating how individual variability in localised concentrations of these molecules relate to brain disorders. Magnetic resonance spectroscopy (MRS) provides a reliable means of measuring, in vivo, concentrations of neurometabolites such as GABA, glutamate and glutamine that can be correlated with brain function and dysfunction. However, an issue of much debate is whether the GABA observed and measured using MRS represents the entire pool of GABA available for measurement (i.e., metabolic, intracellular, and extracellular) or is instead limited to only some portion of it. GABA function can also be investigated indirectly in humans through the use of non-invasive transcranial magnetic stimulation (TMS) techniques that can be used to measure cortical excitability and GABA-mediated physiological inhibition. To investigate this issue further we collected in a single session both types of measurement, i.e., TMS measures of cortical excitability and physiological inhibition and ultra-high-field (7 T) MRS measures of GABA, glutamate and glutamine, from the left sensorimotor cortex of the same group of right-handed individuals. We found that TMS and MRS measures were largely uncorrelated with one another, save for the plateau of the TMS IO curve that was negatively correlated with MRS-Glutamate (Glu) and intra-cortical facilitation (10ms ISI) that was positively associated with MRS-Glutamate concentration. These findings are consistent with the view that the GABA concentrations measured using the MRS largely represent pools of GABA that are linked to tonic rather than phasic inhibition and thus contribute to the inhibitory tone of a brain area rather than GABAergic synaptic transmission.},
+  langid = {english},
+  file = {/Users/tomaubier/Library/Mobile Documents/iCloud~md~obsidian/Documents/UsNeuStim/Files/Dyke2017.pdf}
+}
+
+@article{Yaakub2023,
+  title = {Transcranial Focused Ultrasound-Mediated Neurochemical and Functional Connectivity Changes in Deep Cortical Regions in Humans},
[…]
