### 2025-03-17 #7814 @Kevin-Mattheus-Moerman
- [ ] Please study the above reference check :point_up: and see if you can address any of the reported potential DOI issues. Secondly can you check that you .bib file only contains the references cited, i.e. purge it of uncited works. You can add/amend DOI entries in your `.bib` file, and call `@editorialbot check references` here to check them again.
