### 2025-04-02 #7814 @Tomaubier
@Kevin-Mattheus-Moerman thank you for the update. To be fully transparent, I am currently in the process of completing my PhD and my university requires me to have a minimum number of accepted papers in order to be eligible to defend. In order to anticipate and adjust my publication strategy, I would be interested in getting a rough estimate of my position on the waiting list and the expected delay before being assigned to an editor. Do you have any visibility on this?
Best regards,
Tom
