### 2025-04-15 #7814 @Kevin-Mattheus-Moerman
@Tomaubier thanks for your message. Unfortunately we have only few editors in the medical imaging space (we are actively recruiting new editors to expand in this area), you have been next in line but, so far the other reviews have not been finished yet. For the moment I will assign this submission to myself now to help speed things along. I have some experience with MRI and Ultrasound and should be able to find suitable reviewers. Thanks for your patience so far.
