### 2025-04-26 #7814 @Kevin-Mattheus-Moerman
@plooney, @thewtex, @han-liu, @peterhollender, @mh-skjelvareid, @thorstone25,  @alessandrofelder, @richardkoehler, @Julie-Fabre, @tsbinns, @ntolley, @JuliusWelzel would you be interested in reviewing this submission for [JOSS (the Journal of Open Source Software](https://joss.theoj.org)? The work is titled _CoperniFUS: A flexible Python-based GUI for stereotaxic Focused UltraSound (FUS) experiment planning_.

JOSS reviews take place on GitHub and focus on [the software :floppy_disk: ](https://github.com/Tomaubier/CoperniFUS) as well as [a short paper :scroll: ](https://raw.githubusercontent.com/openjournals/joss-papers/joss.07814/joss.07814/10.21105.joss.07814.pdf). We typically aim for a 2-3 week review but we can be more flexible if needed. You can find more about our review process in our [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).  Here is an example of an ongoing review if you want to see what it is like: https://github.com/openjournals/joss-reviews/issues/7564.

If you are interested you can let me know by responding here.

Thanks!
