### 2025-04-27 #7814 @richardkoehler
I'd be happy to review the paper and software
