### 2025-04-28 #7814 @plooney
I will be a bit busy over the coming weeks but 5-6 weeks is possible for me.

On Mon, 28 Apr 2025, 09:49 Kevin Mattheus Moerman, ***@***.***>
wrote:

> *Kevin-Mattheus-Moerman* left a comment (openjournals/joss-reviews#7814)
> <https://github.com/openjournals/joss-reviews/issues/7814#issuecomment-2834486126>
>
> @plooney <https://github.com/plooney>, @thewtex
> <https://github.com/thewtex>, @han-liu <https://github.com/han-liu>,
> @peterhollender <https://github.com/peterhollender>, @mh-skjelvareid
> <https://github.com/mh-skjelvareid>, @alessandrofelder
> <https://github.com/alessandrofelder>, @tsbinns
> <https://github.com/tsbinns>, @ntolley <https://github.com/ntolley>,
> @JuliusWelzel <https://github.com/JuliusWelzel> would you be interested
> in reviewing this submission for JOSS (the Journal of Open Source Software
> <https://joss.theoj.org>? The work is titled *CoperniFUS: A flexible
> Python-based GUI for stereotaxic Focused UltraSound (FUS) experiment
> planning*.
>
> JOSS reviews take place on GitHub and focus on the software 💾
> <https://github.com/Tomaubier/CoperniFUS> as well as a short paper 📜
> <https://raw.githubusercontent.com/openjournals/joss-papers/joss.07814/joss.07814/10.21105.joss.07814.pdf>.
> We typically aim for a 2-3 week review but we can be more flexible if
> needed. You can find more about our review process in our review
> guidelines
> <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>. Here is
> an example of an ongoing review if you want to see what it is like: #7564
> <https://github.com/openjournals/joss-reviews/issues/7564>.
>
> If you are interested you can let me know by responding here.
>
> Thanks!
>
