### 2025-04-28 #8128 @editorialbot
[REVIEW]: CoperniFUS: A flexible Python-based GUI for stereotaxic Focused UltraSound (FUS) experiment planning
**Submitting author:** @Tomaubier (<a href="http://orcid.org/0009-0004-5558-8435">Tom Aubier</a>)
**Repository:** https://github.com/Tomaubier/CoperniFUS
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.2
