### 2025-05-10 Tomaubier/CoperniFUS#2 @Julie-Fabre
Rendering is not working properly
Hi,

Rendering is not working properly for me in CoperniFUS (see attached screenshot). Could you suggest some ways to debug?
I have blender installed (and rendering inside blender is working).
I am running Ubuntu 24.04 on a Framework Laptop 13 with an Intel® Core™ Ultra 7 165H × 22 processor, Intel® Arc™ Graphics (MTL) and 64GiB of RAM.
I also get the same error on a Windows 11 desktop with 128GB RAM, 13th Gen Intel(R) Core(TM)i7-13700K processor and 12GB NVIDIA GeForce RTX 4070 Graphics card.

![Image](https://github.com/user-attachments/assets/10e09a40-a60b-4c10-87ae-c3b9b8388543)
