### 2025-05-10 #8128 @Julie-Fabre
I haven't gotten the rendering to work properly for me yet, opened [issue#2](https://github.com/Tomaubier/CoperniFUS/issues/2)
