### 2025-05-11 e70e210
disable_internal_console in command line argparse
