### 2025-05-11 37652bc
Dev install procedure in docs
