### 2025-05-11 Tomaubier/CoperniFUS#2 @Tomaubier
Hi,
Thanks for the feedback! Unfortunately I haven’t encountered this issue yet. Blender is only required in this project to perform boolean operations (Blender is one of `trimesh`’s boolean engine). The rendering engine is using `opengl` via `pyqtgraph`.

From your screenshot it seems that the abnormal behavior is only impacting volumetric data rendering. This is handled by `pyqtgraph`’s [GLVolumeItem](https://pyqtgraph.readthedocs.io/en/latest/api_reference/3dgraphics/glvolumeitem.html) method. A quick way to test if `GLVolumeItem` behaves correctly is to run `pyqtgraph`’s Volumetric example.
1. Within the python environnement in which CoperniFUS is installed you can run
   `python -m pyqtgraph.examples`
2. A GUI should pop up, in the `3D graphics` section in the left panel, double click on `Volumetric`
For reference this is the expected output:

If the example runs as expected, the issue has to lie in my code! A `pyqtgraph` or `OpenGL` should be returned in the console.. Just to make sure that `CoperniFUS`’s internal console does not hide any error message from us the latter can be disabled entirely by setting `disable_internal_console` to `True`. When using `CopeniFUS` in standalone mode (launched from a terminal) this option can be set as follows:
`CoperniFUS --disable_internal_console True`
> I’ve just made this argument available from the console, make sure to update CoperniFUS before trying by running
> 1. `pip uninstall CoperniFUS`
> 2. `pip install git+https://github.com/Tomaubier/CoperniFUS.git`

Let me know how this goes.

Best,

Tom
