### 2025-05-12 43a8992
GLVolumeItem input as ubytes
