### 2025-05-12 Tomaubier/CoperniFUS#2 @Tomaubier
Looking back at `pyqtgraph`’s `GLVolumeItem` documentation, it appears that the function is expecting arrays in the `ubytes` type. In my implementation of the BrainAtlas module, I’m providing `GLVolumeItem` with integers ranging from 0 to 255 but still in the `float64` type. For some reason this didn’t cause any issues on my side but some GPUs might not handle this.

I have just pushed an update with this fix, please try it out me let me know how it goes.

Best,

Tom
