### 2025-05-12 82333a5
Cleaning up remaining references to depricated bg-atlasapi -> renamed to brainglobe-atlasapi
