+++ requierements.txt
@@ -2,7 +2,6 @@ PyQt6==6.7.0
 numpy==1.26.4
 scipy==1.13.1
 pyqtgraph==0.13.7
-bg_atlasapi==1.0.3
 brainglobe_atlasapi==2.0.7
 scikit-spatial==7.2.0
 si-prefix==1.3.3
