### 2025-05-13 #8128 @Tomaubier
As a follow up to my response to [issue#3](https://github.com/Tomaubier/CoperniFUS/issues/2), it turns out I only have access to systems running Ubuntu 20.. I spent some time trying to get CoperniFUS to work but I'm still facing errors related to PyQt6. I assume that the occurence of these errors are very system- and version-dependent..
Would you mind trying the installation procedure? If, given your configuration the install completes successfully on the first try you could proceed with the review otherwise we can consider finding an other reviewer.

Let me know what you think.

Best,

Tom
