### 2025-05-13 Tomaubier/CoperniFUS#2 @Julie-Fabre
Hi Tom,

Thanks a lot for the detailed explanations!
To answer your points:
- Testing GLVolumeItem everything looks fine (see below)

![Image](https://github.com/user-attachments/assets/7babb227-4cc2-4f39-9168-7d647d66a661)
- Thanks for the code updates! I still get this error with your most recent push and I don't see any error messages printed in the shell
This is irrespective of OS - I still get errors on both Linux and Windows.

Let me know if you have any more ideas to debug! :)
