### 2025-05-13 Tomaubier/CoperniFUS#2 @Tomaubier
Ok that's good, however things are getting a bit more confusing to me! ^^
As mentioned in [issue#3](https://github.com/Tomaubier/CoperniFUS/issues/3#issuecomment-2877471463) I managed to get everything to run on Ubuntu 24.04. Given the validation I did on windows 11 it doesn't look like it is an OS-related problem..

I'm still trying to think of an intermediate test that could reveal the bug. Out of curiosity, what is happening when you try to pan or rotate the view? Is the viewer responsive? Have you tried changing the resolution of the atlas or the altas type?

> By the way I realize writing this that the resolution selector is not very obvious when getting started with CoperniFUS. I will update the UI to make the existence of this option more obvious because this parameter can drastically change the atlas appearance.
