### 2025-05-14 a8c5c22
anatomical_calib_demo gif in Key features
