### 2025-05-14 33b0004
Linux installation instructions in Github readme
