### 2025-05-14 #8128 @plooney
@Tomaubier There are a number of animal images that I can download. Can you confirm if these comply with the JOSS?

Human and animal research: If the paper contains original data research on humans subjects or animals, does it comply with [JOSS's human participants research policy and/or animal research policy](https://joss.readthedocs.io/en/latest/policies.html?highlight=animal#joss-policies)? If the paper contains no such data, please check this item.
