### 2025-05-14 7f7118b
Force trimesh boolean backend to Blender + error message if blender unavailable
