### 2025-05-14 #8128 @Tomaubier
Brain Atlases accessible in CoperniFUS are the one made available via the [`BrainGlobe Atlas API`](https://brainglobe.info/documentation/brainglobe-atlasapi/index.html). This python package has been published in JOSS back in 2020 ([see 10.21105/joss.02668](https://joss.theoj.org/papers/10.21105/joss.02668)) so I assumed that the journal requirements were satisfied. Nevertheless, I'm going to look at this in more detail, thanks for the heads up.
