### 2025-05-14 Tomaubier/CoperniFUS#2 @Julie-Fabre
Hi Tom,
Frustratingly, I am not sure what combination of re-installing etc did it, but everything seems to be working fine now :)

![Image](https://github.com/user-attachments/assets/d1efc57d-e14b-4096-ab6d-1f0c2eb36dc2)
