### 2025-05-15 b163d3c
Add descriptions to parameter text fields
