### 2025-05-16 Tomaubier/CoperniFUS#6 @Julie-Fabre
JOSS review details
Good job! Overall, I think this software satisfies JOSS requirements: it is novel, responds to a current need in the field, it mostly works and is relatively well documented.
Ideally, a few points still need addressing:
- [x] **Automated tests** The automated software tests appear pretty minimal (unless I missed some of them?). I did not find where they were documented.
- [x] https://github.com/Tomaubier/CoperniFUS/issues/8
- [x] **Functionality**:
    - [x] How to set bregma and lambda are not well documented.
    - [x] How have the skulls been aligned to atlases - in particular the Allen mouse atlas? This Atlas is not in bregma/lambda space but CCF.
    - [x] Some issues need to addressed/features need to be documented - detailed [here](https://github.com/Tomaubier/CoperniFUS/issues/5)

I also have some small additional comments:
- Software:
   - [x] https://github.com/Tomaubier/CoperniFUS/issues/10
- Software documentation:
   - [x] The readthedocs is hard to see in the readMe. My personal preference would be having a link at the top of the readMe.
   - [x] The contribution guidelines aren’t listed in the readthedocs
   - [x] The readthedocs are missing a blurb about what this software is and what is intended for
   - [x] Specify any minimum hardware specifications
- JOSS paper:
   - [x] Kranion seems to be a popular similar tool (although for human planning mainly) - this might also be worth citing? Maybe I missed this.
   - Spelling & grammar:
      - [x] Line 11: "invasive yet targeted alternative to existing neurostimulation modalities" - correct spelling
      - [x] Line 25: "CoperniFUS aims to overcome this obstacle" - correct spelling
      - [x] Line 43-44: "assessment of region-specific responses" - correct spelling
      - [x] Line 78: "to fit a large range of research needs" - grammatically correct
      - [x] Line 136: "can indeed be started as a standalone software" - grammatically correct
