### 2025-05-19 Tomaubier/CoperniFUS#7 @richardkoehler
JOSS Review
Hi there! I think you did a great job on the JOSS paper, but there is some work left to do on the repository side. If you address these issues we should be able to tick all the boxes in the reviewing checklist. Let me know if you have any questions or comments!
Cross-referencing the JOSS issue: https://github.com/openjournals/joss-reviews/issues/8128

### JOSS Paper
- [ ] Maybe **add weblinks** for the trimesh and k-wave packages.
- [ ] **Check the spelling** - overall the paper was well written, but there were some minor spelling errors, that I think should be caught by an automated spell checker.

### Software

- [ ] The **atlas subsampling default could be set to 1**. At first I thought the software was broken until I figured out this parameter was the issue. If you want to keep the default at 10, you must make sure that it is mentioned very clearly in the installation and how-to-use instructions (hoping I didn't overlook anything). Is there a reason this was set at 10?
- [ ] **Add more tests for all the basic functionality**. Currently you only have two very short tests. One codebase with a UI with a good set of tests I am familiar with is MNE, maybe you can draw some inspiration for tests there: https://github.com/mne-tools/mne-python/tree/main/mne/viz/tests or https://github.com/mne-tools/mne-qt-browser/blob/main/mne_qt_browser/tests/test_pg_specific.py
- [ ] Look into **semantic versioning**. Most commonly semantic versioning is used (major, minor, patch), so that your version would be 0.1.0. Since this is one of the most common standards, it would make sense to adhere to it: https://semver.org
- [ ] **pyproject.toml: [project].license**. ``text = `` is deprecated, you can just write e.g. ``license = "GPL-3.0-or-later"``, and move this to the ``[project]`` section: https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#license-and-license-files
- [ ] **pyproject.toml: [project]**. Maybe add the missing keywords in the ``[project]`` that are shown in the full example here, e.g. license-files, keywords, classifiers, readme: https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#a-full-example
- [ ] **pyproject.toml: [build-system]: requires**. You specify ``requires  = ["setuptools >= 69", "setuptools_scm[toml]>=8"]``. I could not find a reason in your codebase why you are using a minimal version of setuptools, and what you are using ``setuptools_scm[toml]`` for (am I overlooking something?). Maybe you just copy-pasted this, or had this for a previous version? If there is no specific reasons for these, you could just write: ``requires = ["setuptools"]`, for me installation still works fine with this.

- [ ] JOSS says the documentation should include a statement of need: **Do the authors clearly state what problems the software is designed to solve and who the target audience is?** I don't feel like this is stated in either the documentation or the README (let me know if I overlooked this). It should probably be stated in both.
- [ ] For a software that is so flexible and extensible, there are **very few examples**. I would recommend to add one or two more examples that showcase this functionality.
- [ ] I could not find **documentation on how to run the tests**. Maybe add this to the CONTRIBUTING file. And add the CONTRIBUTING section to the documentation page as well.
- [ ] Some of the **links in the CONTRIBUTING file are not working**, e.g. [implemented by someone](https://github.com/Tomaubier/CoperniFUS/blob/main/issues#your-first-code-contribution) does not really lead anywhere. And the issues hyperlink address does not even exist: [GitHub issues](https://github.com/Tomaubier/CoperniFUS/blob/main/issues). Make sure to check if all of the links are working, I only tested the 2 and they both didn't work.
- [ ] **The API is only very sparely documented**. Please try to add docstrings with at least a brief description of the function and its parameters and return values. This is of course a bit of work, but I personally find that LLM tools can speed this up quite a lot (for example generate documentation and then proofread, edit, and correct).

### Other minor fixes
- [ ] Missing ``import sys`` in __main__.py
- [ ] Typo: ``Lauching CoperniFUS`` instead of ``Lau**n**ching CoperniFUS`` when starting up the software
- [ ] Typo: ``requi**e**rements.txt`` instead of ``requirements.txt``
