### 2025-05-19 #8128 @Tomaubier
Thank you @Julie-Fabre and @richardkoehler for the feedback! I was a bit in a rush when getting the software ready for submission, it looks like I’ve overlooked basic elements.
I just got shoulder surgery this morning so things are a bit on standby right now but I should hopefully be able to address your comments in the coming days!
Regards,
Tom
