+++ docs/coperniFUS.modules.stereotaxic_frame.rst
@@ -22,7 +22,7 @@ To interact with data from armatures, ``Armature`` objects can be grabbed using 
 
    In [1]: from coperniFUS.viewer import coperniFUSviewer
    In [2]: cfv = coperniFUSviewer()
-      Lauching CoperniFUS
+      Launching CoperniFUS
    In [3]: cfv.stereotaxic_frame.armatures_objects
       {
          'Skull acoustic window': <coperniFUS.modules.armatures.mesh_armatures.STLMeshBooleanArmature at 0x317446900>,
