### 2025-05-20 9d7f6c1
Fixing pyproject depricated architecture
