### 2025-05-20 54c6cad
adding missing references
