+++ joss_paper/paper.md
@@ -31,7 +31,7 @@ bibliography: references.bib
 
 # Summary
 
-Focused UltraSound (FUS) is gaining increasing interest for its potential as a minimally-invasive yet targeted alternative to existing neurostimulation modalities.
+Focused UltraSound (FUS) is gaining increasing interest for its potential as a minimally invasive yet targeted alternative to existing neurostimulation modalities.
 Although reversible changes in the activity of neural structures have been reported as far back as 1929 [@Harvey1929], comprehensive descriptions of the short- and long-term effects of ultrasound on neural structures are still lacking to achieve neurostimulation with satisfactory levels of control and safety. Delivering well characterized FUS pulses with a high degree of spatial selectivity along with local assessment of the state and activity of specific brain regions is crucial in pursuing this research.
 Unlike most electrophysiology procedures involving compact needle-like probes that can be achieved using stereotaxic frame limited to three degrees of freedom, FUS experiments on small _in vivo_ models often require the implementation of complex probe layouts to assess the activity of the stimulated neural structures.
 Treatment planning, evaluation of acoustic parameters through simulations, and post-processing of results often rely on distinct softwares with their own coordinate systems which greatly complicates the integration, analysis and interpretation of key information, thus limiting the ability to properly describe the spatiotemporal aspects of FUS neurostimulation dynamics.
@@ -47,7 +47,7 @@ In an effort to assess the therapeutic potential of ultrasound neurostimulation,
 Studies on rodent models have been pursued using invasive methods such as microdialysis [@Min2011; @Yang2012]. Although these studies report effects of FUS stimulations on Dopamine, Serotonin or GABA levels, their observations are restricted to low ultrasound central frequencies resulting in poorly spatially-selective stimulation, which complicates the assessment of region-specific responses. Despite the interest in characterizing spatially selective stimulations to characterize any region-dependence [@Suarez-Castellanos2021; @Murphy2024a], the choice of low frequencies is typically made to maximize energy transfer through the skull and minimize pressure field distortions. Transducer placement and targeting of the structure is empirical, based on trigonometric evaluation of the focus relative to reference atlases.
 
 <!-- // acoustic simulations -->
-With the growing interest in transcranial ultrasound therapeutic approaches, extensive research has been conducted to develop and validate computational models of acoustic wave propagation through the skull. Although a number of tools and formalisms exist [@Aubry2022a; @Murphy2025], k-Wave [@Treeby2010] has been widely adopted in the field of ultrasound neurostimulation specifically [@Constans2018; @Verhagen2019; @Yaakub2023]. Acoustic simulations in this context are performed using standalone Matlab scripts for the definition of acoustic sources and domains. Acoustic domains are defined either directly based on CT or pseudo-CT scans [@Aubry2003], or by constructing maps from rasterized brain and skull meshes. Registration of the transducer location relative to targeted brain structures is often achieved using frameless optical tracking neuronavigation systems on human or non-human primate subjects [@Murphy2025]. However, empirical methods are usually chosen in small rodents experiments due to the space constraints associated with these models.
[…]
