+++ joss_paper/references.bib
@@ -12,6 +12,46 @@
   file = {/Users/tomaubier/Library/Mobile Documents/iCloud~md~obsidian/Documents/UsNeuStim/Files/Harvey1929.pdf}
 }
 
+@article{Pichardo2023a,
+  title = {{{BabelBrain}}: {{An Open-Source Application}} for {{Prospective Modeling}} of {{Transcranial Focused Ultrasound}} for {{Neuromodulation Applications}}},
+  shorttitle = {{{BabelBrain}}},
+  author = {Pichardo, Samuel},
+  year = {2023},
+  month = jul,
+  journal = {IEEE Transactions on Ultrasonics, Ferroelectrics, and Frequency Control},
+  volume = {70},
+  number = {7},
+  eprint = {2301.12727},
+  primaryclass = {physics},
+  pages = {587--599},
+  issn = {0885-3010, 1525-8955},
+  doi = {10.1109/TUFFC.2023.3274046},
+  urldate = {2025-05-20},
+  abstract = {BabelBrain is an open-source standalone graphic-user-interface application designed for studies of neuromodulation using transcranial focused ultrasound. It calculates the transmitted acoustic field in the brain tissue, taking into account the distortion effects caused by the skull barrier. The simulation is prepared using scans from magnetic resonance imaging (MRI) and, if available, computed tomography and zero-echo time MRI scans. It also calculates the thermal effects based on a given ultrasound regime, such as the total duration of exposure, the duty cycle, and acoustic intensity. The tool is designed to work in tandem with neuronavigation and visualization software, such as 3DSlicer. It uses image processing to prepare domains for ultrasound simulation and uses the BabelViscoFDTD library for transcranial modeling calculations. BabelBrain supports multiple GPU backends, including Metal, OpenCL, and CUDA, and works on all major operating systems including Linux, macOS, and Windows. This tool is particularly optimized for Apple ARM64 systems, which are common in brain imaging research. The paper presents the modeling pipeline used in BabelBrain and a numerical study where different methods of acoustic properties mapping were tested to select the best method that can reproduce the transcranial pressure transmission efficiency reported in the literature.},
+  archiveprefix = {arXiv},
+  langid = {english},
+  keywords = {Physics - Medical Physics},
+  file = {/Users/tomaubier/Zotero/storage/C47PQLKA/Pichardo - 2023 - BabelBrain An Open-Source Application for Prospec.pdf}
+}
+
+@article{Sammartino2020,
+  title = {Kranion, an Open-Source Environment for Planning Transcranial Focused Ultrasound Surgery: Technical Note},
+  shorttitle = {Kranion, an Open-Source Environment for Planning Transcranial Focused Ultrasound Surgery},
+  author = {Sammartino, Francesco and Beam, Dylan W. and Snell, John and Krishna, Vibhor},
+  year = {2020},
+  month = apr,
+  journal = {Journal of Neurosurgery},
+  volume = {132},
+  number = {4},
+  pages = {1249--1255},
+  issn = {0022-3085, 1933-0693},
+  doi = {10.3171/2018.11.JNS181995},
+  urldate = {2025-05-20},
[…]
