### 2025-05-20 f9d14d9
Dynamic atlas subsampliing stride
