### 2025-05-22 df1a8bb
Referencing builtin modules in get_module_object_from_name
