+++ coperniFUS/modules/module_base.py
@@ -9,8 +9,7 @@ class Module:
     def __init__(self, parent_viewer, module_id, **kwargs) -> None:
         self.module_kwargs = kwargs
         self.parent_viewer = parent_viewer
-        self._module_id = None
-        self.module_id = module_id
+        self._module_id = module_id
     
     @property
     def module_id(self):
