### 2025-05-22 7ca23f9
Allow empty atlas selection + atlas coordinate axes ordering correction
