+++ coperniFUS/modules/atlas.py
@@ -1,4 +1,5 @@
 from brainglobe_atlasapi.bg_atlas import BrainGlobeAtlas
+import brainglobe_space as bgs
 import brainglobe_atlasapi
 
 from coperniFUS import *
@@ -6,10 +7,18 @@ from coperniFUS.modules.module_base import Module
 
 class BrainAtlas(Module):
 
+    """
+    BrainAtlas module.
+
+    Attributes:
+        bg_atlas: BrainGlobe Atlas API atlas object.
+    """
+
+    COPERNIFUS_AXES_ORDERING_CONVENTION = "pri" # Postirior Right Inferior see brainglobe_space docs
     _DEFAULT_N_VOXELS = 1e6 # target number of voxels for default atlas subsampling stride computation
     _DEFAULT_PARAMS = {
         'default_atlas_name': 'whs_sd_rat_39um',
-        'highlighted_structure': 'Select structure',
+        'highlighted_structure': 'Select Structure',
         'highlighted_structure_hemisphere': 'Both',
         'atlas_transforms_str' : 'Rx0deg Tz0um',
         'subsampling_stride': 10,
@@ -36,7 +45,7 @@ class BrainAtlas(Module):
         self._brain_atlas_tmat = None
         self.atlas_glvol = None
         self.bg_atlas = None
-        self.bg_atlas_structures = {'Select structure': None}
+        self.bg_atlas_structures = {'Select Structure': None}
 
         self._atlas_voxel_coordinates = None
         self._slicing_plane_mask = None
@@ -97,7 +106,7 @@ class BrainAtlas(Module):
         self.structure_selector = pyqtw.QComboBox()
         # self.structure_selector.setStyleSheet("QComboBox { combobox-popup: 0; }") # Limit dropdown height
         # self.structure_selector.setMaxVisibleItems(20) # Limit dropdown height # BUG darkmode interference
-        self.structure_selector.addItems(['Select structures'])
+        self.structure_selector.addItems(self.bg_atlas_structures.values())
         self.structure_selector.setEnabled(False)
         self.dock_layout.addWidget(self.structure_selector, 1, 0, 1, 1)
         self.structure_selector.setToolTip('Select the brain structure to be highlighted then click on Highlight Structure.<br>You can find detailed descriptions of the available structure on BrainGlobe\'s documentation.')
@@ -112,6 +121,50 @@ class BrainAtlas(Module):
         self.highlight_structure_btn.clicked.connect(self._highlight_structure_btn_pressed)
         self.highlight_structure_btn.setEnabled(False)
         self.dock_layout.addWidget(self.highlight_structure_btn, 1, 2, 1, 1)
+
+        self.update_atlas_selector()
+
+    # ========= TODO brainglobe install -a allen_mouse_25um
+
+    def _add_atlas(self):
+        self.delete_rendered_object()
+        self.init_attributes()
+
+        selected_atlas_description = self.atlas_selector.currentText()
+
+        if selected_atlas_description in ['Select Atlas', 'Offline mode -> Only downloaded atlas are visible']:
+            self.parent_viewer.cache.set_attr('atlas.default_atlas_name', 'no_atlas')
+        elif selected_atlas_description.endswith('(DOWNLOADED)'):
+            offline_atlas_name = selected_atlas_description.split(' | ')[0]
+            self.parent_viewer.statusBar().showMessage(f'Loading offline {offline_atlas_name}', self.parent_viewer._STATUS_BAR_MSG_TIMEOUT)
+            self.bg_atlas = BrainGlobeAtlas(offline_atlas_name, check_latest=False)
+        elif selected_atlas_description.endswith('(online)'):
+            online_atlas_name = selected_atlas_description.split(' | ')[0]
+            
[…]
