### 2025-05-22 bda195f
example mouse atlas as default
