### 2025-05-22 fc3fc08
externalized atlas download
