### 2025-05-22 8df8ff6
Unified atlas space convention + tooltip global loc in statusbar
