+++ coperniFUS/viewer.py
@@ -15,7 +15,7 @@ class Window(pyqtw.QMainWindow):
     y_GREEN = '#7fd169'
     z_BLUE = '#497ccc'
 
-    ATLAS_SPACE_CONVENTION = ['Posterior', 'Left', 'Superior'] # defined according to https://brainglobe.info/documentation/brainglobe-space/index.html
+    ATLAS_SPACE_CONVENTION = ['Anterior', 'Left', 'Superior'] # defined according to https://brainglobe.info/documentation/brainglobe-space/index.html
 
     _STATUS_BAR_MSG_TIMEOUT = 5000
 

[…]
