+++ coperniFUS/modules/anatomical_landmarks_calibration_helper.py
@@ -16,13 +16,26 @@ class AnatLandmarksCalib(Module):
         'cal_tmat': np.eye(4),
     }
 
+    _LANDMARKS_STYLES_BY_SUBTYPE = {
+        'Uncalibrated': {
+            'pt_color': (1, 0, 0, .7), 'pt_size': 18,
+        }, 'Calibrated': {
+            'pt_color': (0, 1, 0, .7), 'pt_size': 16,
+        }
+    }
+
     def __init__(self, parent_viewer, **kwargs) -> None:
         super().__init__(parent_viewer, 'anat_calib', **kwargs)
 
-        self.anat_landmarks_dict_name2ref = {
+        self._anat_landmarks_dict_name2ref = {
             'Uncalibrated': 'uncal_anatomical_landmarks_coords',
             'Calibrated': 'cal_anatomical_landmarks_coords',
         }
+        self._anat_landmarks_dict_ref2name = {
+            ref: name for name, ref in self._anat_landmarks_dict_name2ref.items()
+        }
+
+        self._tooltip_to_landmark_full_name = None
         self._landmarks_calib_tmat = None
         self._landmarks_gl_items = {}
 
@@ -37,62 +50,300 @@ class AnatLandmarksCalib(Module):
         self.dock_layout = pyqtw.QGridLayout()
         self.dock_widget.setLayout(self.dock_layout)
 
-        # Landmark to tooltip selector
-        self.landmark_selector = pyqtw.QComboBox()
-        self.update_landmark_selector_elements()
-        self.dock_layout.addWidget(self.landmark_selector, 0, 0, 1, 1) # Y, X, w, h
+        # Tree view and model
+        self.tree_view = DynamicallyResizableTreeView()
+        self.tree_view.resized.connect(self._on_tree_view_resize)
+        self.model = pyqtg.QStandardItemModel()
+        self.model.setHorizontalHeaderLabels(["Anatomical Landmarks", "Tooltip"])
+        self.tree_view.setModel(self.model)
+        self.dock_layout.addWidget(self.tree_view, 0, 0, 1, 3) # Y, X, h, w
+        self._populate_treeview_cached_landmarks()
 
-        # Landmark to tooltip button
-        self.landmark_to_tooltip_btn = pyqtw.QPushButton('Set Landmark to Tooltip')
-        self.landmark_to_tooltip_btn.clicked.connect(self.set_landmark_to_tooltip)
-        self.dock_layout.addWidget(self.landmark_to_tooltip_btn, 0, 1, 1, 1) # Y, X, w, h
+        # Add / Remove Landmarks Buttons
+        self.add_landmark_button = pyqtw.QPushButton('+')
+        self.add_landmark_button.setFixedWidth(25)
+        self.dock_layout.addWidget(self.add_landmark_button, 1, 0, 1, 1) # Y, X, h, w
+        self.add_landmark_button.clicked.connect(self._add_landmark_btn_pressed)
+        self.add_landmark_button.setToolTip('Add an Anatomical Landmark. Landmarks names can be edited later in the list.')
+        
+        self.remove_landmark_button = pyqtw.QPushButton('-')
+        self.remove_landmark_button.setFixedWidth(25)
+        self.dock_layout.addWidget(self.remove_landmark_button, 1, 1, 1, 1) # Y, X, h, w
+        self.remove_landmark_button.clicked.connect(self._remove_landmark_btn_pressed)
+        self.remove_landmark_button.setToolTip('Remove the selected Anatomical Landmark.')
+
+        # Connect model item changed signal
+        self.model.itemChanged.connect(self._handle_landmark_item_change)
+
+        # Set coords btn
[…]
