### 2025-06-02 386714b
Transform matrices validity check in setter methods
