+++ coperniFUS/modules/atlas.py
@@ -656,15 +656,15 @@ class BrainAtlas(Module):
             target_space = bgs.AnatomicalSpace(get_flipped_atlas_space_convention(self.parent_viewer.ATLAS_SPACE_CONVENTION))
             axes_order_idx = source_space.map_to(target_space)[0]
             space_conversion_tmat = source_space.transformation_matrix_to(target_space)
-            self.brain_atlas_tmat = space_conversion_tmat.T
+            self._brain_atlas_tmat = space_conversion_tmat.T
 
             # Setting atlas scale based on resolution
-            self.brain_atlas_tmat = self.brain_atlas_tmat @ af_tr.scale_mat(resolution)
+            self._brain_atlas_tmat = self._brain_atlas_tmat @ af_tr.scale_mat(resolution)
 
             # Img space origin to atlas center
-            self.brain_atlas_tmat = self.brain_atlas_tmat @ af_tr.translat_mat('x', -(resolution[0] * atlas_shape[axes_order_idx[0]]) / 2)
-            self.brain_atlas_tmat = self.brain_atlas_tmat @ af_tr.translat_mat('y', -(resolution[1] * atlas_shape[axes_order_idx[1]]) / 2)
-            self.brain_atlas_tmat = self.brain_atlas_tmat @ af_tr.translat_mat('z', -(resolution[2] * atlas_shape[axes_order_idx[2]]) / 2)
+            self._brain_atlas_tmat = self._brain_atlas_tmat @ af_tr.translat_mat('x', -(resolution[0] * atlas_shape[axes_order_idx[0]]) / 2)
+            self._brain_atlas_tmat = self._brain_atlas_tmat @ af_tr.translat_mat('y', -(resolution[1] * atlas_shape[axes_order_idx[1]]) / 2)
+            self._brain_atlas_tmat = self._brain_atlas_tmat @ af_tr.translat_mat('z', -(resolution[2] * atlas_shape[axes_order_idx[2]]) / 2)
 
             atlas_transforms_matrices = af_tr_from_str.transform_matrices_from_str(
                 self.get_user_param('atlas_transforms_str')
