### 2025-06-02 f91a443
Bumping trimesh to 4.6.10 (blender is no longer necessary for boolean operations -> manifold backend) + k-wave-python to 0.4.0
