+++ docs/requirements.txt
@@ -7,14 +7,14 @@ sphinx_mdinclude==0.6.2
 
 PyQt6==6.7.0
 numpy==1.26.4
-scipy==1.13.1
+scipy==1.14.1
+networkx==3.5
 pyqtgraph==0.13.7
 brainglobe_atlasapi==2.0.7
 scikit-spatial==7.2.0
 si-prefix==1.3.3
 numpy-stl==3.1.1
-trimesh==4.0.1
-shapely==2.0.5
+shapely==2.1.1
 rtree==1.3.0
 mapbox_earcut==1.0.1
 zarr==2.18.3
@@ -23,7 +23,10 @@ tqdm==4.66.4
 dask==2025.3.0
 rembg==2.0.59
 opencv-python==4.10.0.84
-matplotlib==3.9.0
-napari==0.5.4
+pymeshfix==0.17.1
+manifold3d==3.1.1
+trimesh==4.6.10
+matplotlib==3.10.0
+napari==0.6.1
 onnxruntime==1.19.2
-k-Wave-python==0.3.4
+k-Wave-python==0.4.0

[…]
