### 2025-06-02 698a567
Brain Atlas Module error popup refactoring
