### 2025-06-02 e49c593
Anat calib module error handling
