+++ tests/test_viewer.py
@@ -3,10 +3,15 @@ from coperniFUS.viewer import Window, pyqtw
 
 @pytest.fixture
 def viewer_window(qtbot):
-    """Fixture to create CoperniFUS viewer window."""
+    """ Fixture to create CoperniFUS viewer window. """
     window = Window(app=None)
     return window
 
 def test_brain_atlas(viewer_window):
-    """Test tha the example atlas has been loaded."""
-    assert viewer_window.get_module_object_from_name('BrainAtlas').bg_atlas.atlas_name == 'example_mouse_100um'
+    """ Test that the example mouse atlas has been loaded."""
+    batlas_module = viewer_window.get_module_object_from_name('BrainAtlas') # Get BrainAtlasModule handle
+    batlas_module.add_reference_atlas('example_mouse_100um') # Loading demo atlas
+    batlas_module.add_structure_layer(structure='Cerebrum (CH)', hemisphere='Left Hemisphere') # Loading a brain structure
+    assert len(batlas_module.layers) == 2
+
+# TODO implement more
