+++ .github/workflows/python-app.yml
@@ -12,7 +12,7 @@ jobs:
         
     strategy:
       matrix:
-        os: [macos-latest, windows-latest] # ubuntu-latest
+        os: [macos-latest, windows-latest, ubuntu-latest] # ubuntu-latest
         python-version: ['3.12']
 
     steps:
@@ -35,19 +35,24 @@ jobs:
         if: runner.os == 'Linux'
         run: |
           sudo apt-get update -y --allow-releaseinfo-change
-          sudo apt-get install --no-install-recommends -y \
-            libxkbcommon-x11-0 \
-            x11-utils \
-            libyaml-dev \
-            libegl1 \
-            libxcb-icccm4 \
-            libxcb-image0 \
-            libxcb-keysyms1 \
-            libxcb-randr0 \
-            libxcb-render-util0 \
-            libxcb-xinerama0 \
-            libopengl0 \
-            libxcb-cursor0
+          sudo apt-get install -y libxcb-cursor-dev
+
+      # - name: "Install Linux VirtualDisplay"
+      #   if: runner.os == 'Linux'
+      #   run: |
+      #     sudo apt-get install --no-install-recommends -y \
+      #       libxkbcommon-x11-0 \
+      #       x11-utils \
+      #       libyaml-dev \
+      #       libegl1 \
+      #       libxcb-icccm4 \
+      #       libxcb-image0 \
+      #       libxcb-keysyms1 \
+      #       libxcb-randr0 \
+      #       libxcb-render-util0 \
+      #       libxcb-xinerama0 \
+      #       libopengl0 \
+      #       libxcb-cursor0
 
       - name: Installing CoperniFUS
         run: python -m pip install -e .
