### 2025-06-02 dae22f4
GH tests ubuntu virtual display
