+++ .github/workflows/python-app.yml
@@ -43,8 +43,8 @@ jobs:
           sudo apt-get update
           sudo apt-get install -y \
             libegl1 \
+            libgl1 \
             xvfb \
-            libgl1-mesa-glx \
             libxkbcommon-x11-0 \
             libxcb-xinerama0
 
