### 2025-06-02 6c666d3
GH action virtual display debug
