+++ .github/workflows/python-app.yml
@@ -75,7 +75,7 @@ jobs:
       #   run: python -m pip install -e .
 
       - name: Installing pqtg
-        run: pip install pyqtgraph
+        run: pip install PyQt6 pyqtgraph
 
       - name: Installing test dependencies
         run: pip install pytest pytest-qt
