+++ .github/workflows/python-app.yml
@@ -31,13 +31,13 @@ jobs:
           7z x mesa.7z -o*
           powershell.exe mesa\systemwidedeploy.cmd 1
       
-      - name: "Install Linux VirtualDisplay"
+      - name: "Install Linux system dependencies"
         if: runner.os == 'Linux'
         run: |
-          sudo apt-get update -y --allow-releaseinfo-change
+          sudo apt-get update
           sudo apt-get install -y libxcb-cursor-dev
 
-      - name: Install system dependencies
+      - name: "Install Linux VirtualDisplay"
         if: runner.os == 'Linux'
         run: |
           sudo apt-get update
@@ -82,11 +82,11 @@ jobs:
       #       libopengl0 \
       #       libxcb-cursor0
 
-      # - name: Installing CoperniFUS
-      #   run: python -m pip install -e .
+      - name: Installing CoperniFUS
+        run: python -m pip install -e .
 
-      - name: Installing pqtg
-        run: pip install PyQt6 pyqtgraph
+      # - name: Installing pqtg
+      #   run: pip install PyQt6 pyqtgraph
 
       - name: Installing test dependencies
         run: pip install pytest pytest-qt
@@ -98,21 +98,21 @@ jobs:
           python -m pyqtgraph.util.glinfo
         # python -m pyqtgraph.util.get_resolution
 
-      # - name: "Debug Info"
-      #   run: |
-      #     echo python location: `which python`
-      #     echo python version: `python --version`
-      #     echo pytest location: `which pytest`
-      #     echo installed packages
-      #     python -m pip list
-      #     echo pyqtgraph system info
-      #     python -c "import pyqtgraph as pg; pg.systemInfo()"
-      #   shell: bash
-      #   env:
-      #     QT_DEBUG_PLUGINS: 1
+      - name: "Debug Info"
+        run: |
+          echo python location: `which python`
+          echo python version: `python --version`
+          echo pytest location: `which pytest`
+          echo installed packages
+          python -m pip list
+          echo pyqtgraph system info
+          python -c "import pyqtgraph as pg; pg.systemInfo()"
+        shell: bash
+        env:
+          QT_DEBUG_PLUGINS: 1
 
-      # - name: Download test BrainGlobe atlas
-      #   run: brainglobe install -a example_mouse_100um
+      - name: Download test BrainGlobe atlas
+        run: brainglobe install -a example_mouse_100um
 
-      # - name: Testing CoperniFUS
-      #   run: pytest tests/test_viewer.py
+      - name: Testing CoperniFUS
+        run: pytest tests/test_viewer.py
