+++ .github/workflows/python-app.yml
@@ -82,11 +82,11 @@ jobs:
       #       libopengl0 \
       #       libxcb-cursor0
 
-      # - name: Installing CoperniFUS
-      #   run: python -m pip install -e .
+      - name: Installing CoperniFUS
+        run: python -m pip install -e .
 
-      - name: Installing pqtg
-        run: pip install PyQt6 pyqtgraph
+      # - name: Installing pqtg
+      #   run: pip install PyQt6 pyqtgraph
 
       - name: Installing test dependencies
         run: pip install pytest pytest-qt
