+++ .github/workflows/python-app.yml
@@ -127,8 +127,15 @@ jobs:
       - name: Download test BrainGlobe atlas
         run: brainglobe install -a example_mouse_100um
 
-      - name: "Testing CoperniFUS (Windows/MacOS)"
-        run: pytest tests/test_viewer.py
+      # - name: "Testing CoperniFUS (Windows/MacOS)"
+      #   run: pytest tests/test_viewer.py
+
+      - name: Run PyQt code headlessly
+        env:
+          QT_QPA_PLATFORM: offscreen
+        run: |
+          python -m pyqtgraph.util.glinfo
+          pytest tests/test_viewer.py
 
       # - name: "Testing CoperniFUS (Linux)"
       #   if: runner.os == 'Linux'
