+++ .github/workflows/python-app.yml
@@ -12,7 +12,7 @@ jobs:
         
     strategy:
       matrix:
-        os: [ubuntu-latest] # [macos-latest, windows-latest] # 
+        os: [macos-latest, windows-latest, ubuntu-latest]
         python-version: ['3.12']
 
     steps:
@@ -59,29 +59,6 @@ jobs:
             libxi6 \
             libxtst6
 
-      # - name: "XVFB Display Info"
-      #   if: runner.os == 'Linux'
-      #   run: |
-      #     xvfb-run --server-args="-screen 0, 1920x1200x24 -ac +extension GLX +render -noreset" python -m pyqtgraph.util.glinfo
-      #     xvfb-run --server-args="-screen 0, 1920x1200x24 -ac +extension GLX +render -noreset" python -m pyqtgraph.util.get_resolution
-
-      # - name: "Install Linux VirtualDisplay"
-      #   if: runner.os == 'Linux'
-      #   run: |
-      #     sudo apt-get install --no-install-recommends -y \
-      #       libxkbcommon-x11-0 \
-      #       x11-utils \
-      #       libyaml-dev \
-      #       libegl1 \
-      #       libxcb-icccm4 \
-      #       libxcb-image0 \
-      #       libxcb-keysyms1 \
-      #       libxcb-randr0 \
-      #       libxcb-render-util0 \
-      #       libxcb-xinerama0 \
-      #       libopengl0 \
-      #       libxcb-cursor0
-
       - name: Installing CoperniFUS
         run: python -m pip install -e .
 
@@ -105,41 +82,30 @@ jobs:
           python -m pyqtgraph.util.glinfo
           python -m pyqtgraph.util.get_resolution
 
-      # - name: "Display Info 2"
-      #   run: |
-      #     export QT_QPA_PLATFORM=offscreen
-      #     xvfb-run --auto-servernum --server-args="-screen 0 1920x1200x24" \
-      #     python -c "from PyQt6.QtGui import QGuiApplication; app = QGuiApplication([]); print(QGuiApplication.platformName())"
-
-      # - name: "Debug Info"
-      #   run: |
-      #     echo python location: `which python`
-      #     echo python version: `python --version`
-      #     echo pytest location: `which pytest`
-      #     echo installed packages
-      #     python -m pip list
-      #     echo pyqtgraph system info
-      #     python -c "import pyqtgraph as pg; pg.systemInfo()"
-      #   shell: bash
-      #   env:
-      #     QT_DEBUG_PLUGINS: 1
+      - name: "Debug Info"
+        run: |
+          echo python location: `which python`
+          echo python version: `python --version`
+          echo pytest location: `which pytest`
+          echo installed packages
+          python -m pip list
+          echo pyqtgraph system info
+          python -c "import pyqtgraph as pg; pg.systemInfo()"
+        shell: bash
+        env:
+          QT_DEBUG_PLUGINS: 1
 
       - name: Download test BrainGlobe atlas
         run: brainglobe install -a example_mouse_100um
 
-      # - name: "Testing CoperniFUS (Windows/MacOS)"
-      #   run: pytest tests/test_viewer.py
+      - name: "Testing CoperniFUS (Windows/MacOS)"
+        if: runner.os != 'Linux'
+        run: pytest tests/test_viewer.py
 
-      - name: Run PyQt code headlessly
+      - name: "Testing CoperniFUS (Linux)"
+        if: runner.os == 'Linux'
         run: |
           xvfb-run --auto-servernum --server-args="-screen 0 1920x1200x24" \
             python -m pyqtgraph.util.glinfo
           xvfb-run --auto-servernum --server-args="-screen 0 1920x1200x24" \
[…]
