### 2025-06-03 beadd8e
Docstring cleanup
