+++ coperniFUS/modules/img_as_plane.py
@@ -13,6 +13,7 @@ class RefImageAsPlane(Module):
         'origin_px_xloc': 0, # [px]
         'origin_px_yloc': 0, # [px]
     }
+    """ Default configuration parameters used when a parameter value is not yet cached """
 
     def __init__(self, parent_viewer, **kwargs) -> None:
         super().__init__(parent_viewer, 'ref_image_as_plane_params', **kwargs)
@@ -22,28 +23,66 @@ class RefImageAsPlane(Module):
         self.ref_image_name = None
         self.ref_image = None
     
-    # --- Atlas specific cache wrapper ---
-    
-    def get_user_param(self, param_name, default_value=None):
-        if self.ref_image_name is not None:
-            param_value = super().get_user_param(
-                param_name,
-                additional_identifiers=[self.ref_image_name],
-                default_value=default_value)
-        else:
-            param_value = self._DEFAULT_PARAMS[param_name]
-        return param_value
+    # --- Module specific public attributes ---
 
-    def set_user_param(self, param_name, param_value):
-        if self.ref_image_name is not None:
-            super().set_user_param(
-                param_name,
-                additional_identifiers=[self.ref_image_name],
-                param_value=param_value)
+    def import_image(self, image_path):
+        """ Import a reference image to the viewer by providing the image_path.
+            Loaded image can be removed by calling the delete_rendered_object method """
+        if image_path == '':
+            raise ValueError('Invalid file path')
+        else:
+            if self._load_img(image_path):
+                self.ref_image_name = pathlib.Path(image_path).stem
+                self.parent_viewer.cache.set_attr(
+                    ['ref_image_as_plane_params', 'last_used_img_name'],
+                    self.ref_image_name
+                )
+                self.set_user_param('file_path', image_path)
+
+    @property
+    def ref_image_tmat(self):
+        """ Holds the reference image affine transformation matrix """
+        if self._ref_image_tmat is None:
+            px_size = self.get_user_param('px_size')
+            self._ref_image_tmat = af_tr.scale_mat(px_size)
+
+            self._ref_image_tmat = self._ref_image_tmat @ af_tr.translat_mat('x', -self.get_user_param('origin_px_xloc') * px_size)
+            self._ref_image_tmat = self._ref_image_tmat @ af_tr.translat_mat('y', -self.get_user_param('origin_px_yloc') * px_size)
+
+            if self.get_user_param('plane') == 'X':
+                self._ref_image_tmat = self._ref_image_tmat @ af_tr.rot_mat('y', 90)
+            if self.get_user_param('plane') == '-X':
+                self._ref_image_tmat = self._ref_image_tmat @ af_tr.rot_mat('y', 90)
+                self._ref_image_tmat = self._ref_image_tmat @ af_tr.rot_mat('z', 180)
+            if self.get_user_param('plane') == 'Y':
+                self._ref_image_tmat = self._ref_image_tmat @ af_tr.rot_mat('y', 90)
+                self._ref_image_tmat = self._ref_image_tmat @ af_tr.rot_mat('z', 90)
+            if self.get_user_param('plane') == '-Y':
+                self._ref_image_tmat = self._ref_image_tmat @ af_tr.rot_mat('y', 90)
+                self._ref_image_tmat = self._ref_image_tmat @ af_tr.rot_mat('z', 90)
+                self._ref_image_tmat = self._ref_image_tmat @ af_tr.rot_mat('z', 180)
+            if self.get_user_param('plane') == '-Z':
+                self._ref_image_tmat = self._ref_image_tmat @ af_tr.rot_mat('x', 180)
+        
[…]
