+++ coperniFUS/modules/tooltip.py
@@ -8,6 +8,7 @@ class Tooltip(Module):
         'tooltip_transforms_str' : 'Rx0deg Tz0um',
         'axes_length': 1e-3
     }
+    """ Default configuration parameters used when a parameter value is not yet cached """
 
     def __init__(self, parent_viewer, **kwargs) -> None:
         super().__init__(parent_viewer, 'tooltip', **kwargs)
@@ -17,9 +18,52 @@ class Tooltip(Module):
         self.y_glaxis = None
         self.z_glaxis = None
 
+    # --- Module specific public attributes ---
+
+    def release_from_modules(self, sender_module=None):
+        """ All modules allowing Tooltip capture have to implement the release_tooltip method """
+        # Builtin modules
+        if sender_module is not self.parent_viewer.stereotaxic_frame:
+            self.parent_viewer.stereotaxic_frame.release_tooltip()
+        if sender_module is not self.parent_viewer.anat_calib:
+            self.parent_viewer.anat_calib.release_tooltip()
+        # Optionnal modules
+        for mm in self.parent_viewer._modules:
+            if hasattr(mm, 'release_tooltip') and sender_module is not mm:
+                mm.release_tooltip()
+
+    @property
+    def tooltip_coordinates(self):
+        """ Returns the current Tooltip 3D coordinates (x, y, z) """
+        _tooltip_coordinates = None
+        if self.tooltip_tmat is not None:
+            _tooltip_coordinates = self.tooltip_tmat[3, :3]
+        return _tooltip_coordinates
+
+    @property
+    def tooltip_tmat(self):
+        """ Holds the current Tooltip affine transformation matrix """
+        if self._tooltip_tmat is None:
+            # Compute tooltip transform matrix from tooltip_transforms_str (status bar qedit)
+            transforms_matrices = af_tr_from_str.transform_matrices_from_str(
+                self.get_user_param('tooltip_transforms_str')
+            )
+            self._tooltip_tmat = af_tr.scale_mat(1)
+            for tr_mat in transforms_matrices:
+                self._tooltip_tmat = self._tooltip_tmat @ tr_mat
+        return self._tooltip_tmat
+    
+    @tooltip_tmat.setter
+    def tooltip_tmat(self, value):
+        if value is not None:
+            if value.shape != (4, 4):
+                raise ValueError('Transformation matrix should be of shape (4, 4)')
+        self._tooltip_tmat = value
+
     # --- Required module attributes ---
 
     def add_rendered_object(self):
+        """ Called when populating the viewer with the module rendered objects """
         self.x_glaxis = gl.GLLinePlotItem(pos=[[0,0,0], [self.get_user_param('axes_length'),0,0]], width=6, color=self.parent_viewer.x_RED, antialias=True, glOptions='additive')
         self.y_glaxis = gl.GLLinePlotItem(pos=[[0,0,0], [0,self.get_user_param('axes_length'),0]], width=6, color=self.parent_viewer.y_GREEN, antialias=True, glOptions='additive')
         self.z_glaxis = gl.GLLinePlotItem(pos=[[0,0,0], [0,0,self.get_user_param('axes_length')]], width=6, color=self.parent_viewer.z_BLUE, antialias=True, glOptions='additive')
@@ -35,10 +79,12 @@ class Tooltip(Module):
         self._update_transform()
 
     def update_rendered_object(self):
+        """ Called on render view updates """
         if self.x_glaxis is not None:
             self._update_transform()
 
     def delete_rendered_object(self):
+        """ Called on deletion of the module rendered objects """
         if self.x_glaxis in self.parent_viewer.gl_view.items:
[…]
