+++ docs/requirements.txt
@@ -30,3 +30,4 @@ matplotlib==3.10.0
 napari==0.6.1
 onnxruntime==1.19.2
 k-Wave-python==0.4.0
+..
