+++ docs/coperniFUS.modules.tooltip.rst
@@ -1,7 +1,7 @@
 Tooltip ``Module``
 ------------------
 
-Allows the user to input coordinates using the :ref:`transformations as string syntax <transformation_strings_syntax>`:, visualized in the viewer by an axis triad.
+Allows the user to input coordinates using the :ref:`transformations as string syntax <transformation_strings_syntax>`, visualized in the viewer by an axis triad.
 Tooltip location can inherit from any stereotaxic frame `Armatures <coperniFUS.modules.stereotaxic_frame.rst>`_ by checking its associated ``tooltip`` checkbox in the Stereotaxic Frame module dock.
 
 Tooltip coordinates can be referenced by any modules, facilitating anatomical landmark registration, or assessment of distances / locations in any coordinate frames.
