+++ coperniFUS/modules/stereotaxic_frame.py
@@ -48,6 +48,81 @@ class StereotaxicFrame(Module):
         """ Returns armatures objects loaded in the stereotaxic module """
         return self._armatures_objects
     
+    def add_armature(self, new_armature_class, new_armature_display_name):
+        """ Call to add an Armature of type new_armature_class to the StereataxicFrame module (provide the class name as a string).
+        The list of available armature types can be obtained in AVAILABLE_ARMATURES.
+        Please provide an armature name in new_armature_display_name.
+        """
+        if new_armature_display_name in self.armatures_objects:
+            raise ValueError(f'{new_armature_display_name} already exist in armatures_objects')
+
+        # --- Add armature to _armatures_objects
+        armature_obj = eval(new_armature_class)(
+            armature_display_name=new_armature_display_name,
+            parent_viewer=self.parent_viewer,
+            stereotax_frame_instance=self)
+        self.armatures_objects[new_armature_display_name] = armature_obj
+
+        # --- Add to stacked parameter editor widgets ---
+        self.armature_parameters_stacked_widget.addWidget(
+            armature_obj.params_editor_widget.armature_params_editor_widget)
+
+        # --- Add armature to _steframe_armatures_objects_clsnames ---
+        armatures_objects_clsnames = self.get_user_param('_steframe_armatures_objects_clsnames')
+        armatures_objects_clsnames[new_armature_display_name] = new_armature_class
+        self.set_user_param('_steframe_armatures_objects_clsnames', armatures_objects_clsnames)
+
+        # --- Add armature to _steframe_arch_dict ---
+        arch_dict = self.get_user_param('_steframe_arch_dict')
+        arch_dict[new_armature_display_name] = None
+        self.set_user_param('_steframe_arch_dict', arch_dict)
+
+        self._update_armatures_qtree()
+
+    def delete_armature(self, armature_name):
+        """ Call to delete 'armature_name' from the StereataxicFrame module. """
+        if not armature_name in self.armatures_objects:
+            raise ValueError(f'{armature_name} does not exist in armatures_objects')
+
+        # --- Pop armature from _steframe_arch_dict ---
+        def pop_nested_key(d, key, popped_armature_childs=None):
+            if isinstance(d, dict):
+                if key in d:
+                    popped_armature_childs = d[key]
+                    d.pop(key, None)
+                for k, v in d.items():
+                    popped_armature_childs = pop_nested_key(v, key, popped_armature_childs)
+            return popped_armature_childs
+
+        arch_dict = self.get_user_param('_steframe_arch_dict')
+        popped_armature_childs = pop_nested_key(arch_dict, armature_name)
+        if isinstance(popped_armature_childs, dict):
+            arch_dict = {**arch_dict, **popped_armature_childs} # Add childs to the end of dict
+        self.set_user_param('_steframe_arch_dict', arch_dict)
+
+        # --- Pop armature from _steframe_armatures_objects_clsnames ---
+        armatures_objects_clsnames = self.get_user_param('_steframe_armatures_objects_clsnames')
+        if armature_name in armatures_objects_clsnames:
+            armatures_objects_clsnames.pop(armature_name)
+        self.set_user_param('_steframe_armatures_objects_clsnames', armatures_objects_clsnames)
+
+        # --- Remove from stacked parameter editor widgets ---
+        try:
+            popped_armature_index = list(self.armatures_objects.keys()).index(armature_name)
+            widget_to_remove = self.armature_parameters_stacked_widget.widget(popped_armature_index)
+            self.armature_parameters_stacked_widget.removeWidget(widget_to_remove)
+            widget_to_remove.deleteLater()
+            self.armature_parameters_stacked_widget.setCurrentIndex(0)
+        except ValueError as e:
[…]
