+++ coperniFUS/modules/interfaces/trimesh_interfaces.py
@@ -84,7 +84,7 @@ class TrimeshHandler:
             self._stl_item_tmat = np.eye(4) # No transform as default
         
         # Apply anatomical landmark calibration transformation if enabled
-        if not self.get_stl_user_param('ignore_anatomical_landmarks_calibration'):
+        if not self.get_user_param('ignore_anatomical_landmarks_calibration'):
             anatomically_calibrated_stl_item_tmat = self._stl_item_tmat @ self.parent_viewer.anat_calib.landmarks_calib_tmat
         else:
             anatomically_calibrated_stl_item_tmat = self._stl_item_tmat
@@ -108,13 +108,13 @@ class TrimeshHandler:
             self.stl_glitem.append(
                 gl.GLMeshItem(
                     meshdata=stl_item_gl_mesh_data,
-                    shader=self.get_stl_user_param('gl_mesh_shader'), # TODO get_stl_user_param -> redondant -> transfer to stl dock
-                    smooth=self.get_stl_user_param('gl_mesh_smooth'),
-                    drawFaces=self.get_stl_user_param('gl_mesh_drawFaces'),
-                    drawEdges=self.get_stl_user_param('gl_mesh_drawEdges'),
-                    edgeColor=self.get_stl_user_param('gl_mesh_edgeColor'),
-                    edgeWidth=self.get_stl_user_param('gl_mesh_edgeWidth'),
-                    glOptions=self.get_stl_user_param('gl_mesh_glOptions'),
+                    shader=self.get_user_param('gl_mesh_shader'), # TODO get_user_param -> redondant -> transfer to stl dock
+                    smooth=self.get_user_param('gl_mesh_smooth'),
+                    drawFaces=self.get_user_param('gl_mesh_drawFaces'),
+                    drawEdges=self.get_user_param('gl_mesh_drawEdges'),
+                    edgeColor=self.get_user_param('gl_mesh_edgeColor'),
+                    edgeWidth=self.get_user_param('gl_mesh_edgeWidth'),
+                    glOptions=self.get_user_param('gl_mesh_glOptions'),
             ))
             bool_mesh_index_str = f' {mesh.bool_mesh_index}' if hasattr(mesh, 'bool_mesh_index') else ''
             self.parent_viewer.gl_view.addItem(self.stl_glitem[-1], name=f'{self.stl_item_name}{bool_mesh_index_str} STL mesh')
@@ -143,7 +143,7 @@ class TrimeshHandler:
             if sub_mesh_index < len(self.stl_glitem):
                 self.stl_glitem[sub_mesh_index].setMeshData(meshdata=stl_item_gl_mesh_data)
 
-        ignore_plane_slicing = self.get_stl_user_param('ignore_plane_slicing')
+        ignore_plane_slicing = self.get_user_param('ignore_plane_slicing')
         if not self.parent_viewer._postpone_slicing_plane_computation or ignore_plane_slicing:
             if self.stl_glitem is None:
                 self.add_rendered_object()
@@ -169,7 +169,7 @@ class TrimeshHandler:
 
     # --- cache wrapper for interface parameters ---
 
-    def get_stl_user_param(self, param_name, default_value=None):
+    def get_user_param(self, param_name, default_value=None):
         """ Get module configuration parameter stored in cache (or default values if non existant) """
         if default_value is None and param_name in self._DEFAULT_PARAMS:
             default_value = self._DEFAULT_PARAMS[param_name]
@@ -182,7 +182,7 @@ class TrimeshHandler:
             param_value = default_value
         return param_value
 
-    def set_stl_user_param(self, param_name, param_value):
+    def set_user_param(self, param_name, param_value):
         """ Set module configuration parameter to cache """
         if self.stl_item_name is not None:
             self.parent_viewer.cache.set_attr(
[…]
