### 2025-06-05 4188517
Kwave test outcome check update
