### 2025-06-05 dca7860
kwave test tolerance update
