### 2025-06-05 7b3c380
macos kwave install debug
