### 2025-06-05 9e5a618
Contributing section in docs
