+++ joss_paper/paper.md
@@ -47,7 +47,7 @@ In an effort to assess the therapeutic potential of ultrasound neurostimulation,
 Studies on rodent models have been pursued using invasive methods such as microdialysis [@Min2011; @Yang2012]. Although these studies report effects of FUS stimulations on Dopamine, Serotonin or GABA levels, their observations are restricted to low ultrasound central frequencies resulting in poorly spatially-selective stimulation, which complicates the assessment of region-specific responses. Despite the interest in characterizing spatially selective stimulations to characterize any region-dependence [@Suarez-Castellanos2021; @Murphy2024a], the choice of low frequencies is typically made to maximize energy transfer through the skull and minimize pressure field distortions. Transducer placement and targeting of the structure is empirical, based on trigonometric evaluation of the focus relative to reference atlases.
 
 <!-- // acoustic simulations -->
-With the growing interest in transcranial ultrasound therapeutic approaches, extensive research has been conducted to develop and validate computational models of acoustic wave propagation through the skull. Although a number of tools and formalisms exist [@Aubry2022a; @Murphy2025], k-Wave [@Treeby2010] has been widely adopted in the field of ultrasound neurostimulation specifically [@Constans2018; @Verhagen2019; @Yaakub2023]. Acoustic simulations in this context are performed using standalone Matlab scripts for the definition of acoustic sources and domains. Acoustic domains are defined either directly based on CT or pseudo-CT scans [@Aubry2003], or by constructing maps from rasterized brain and skull meshes. Registration of the transducer location relative to targeted brain structures is often achieved using frameless optical tracking neuronavigation systems on human or non-human primate subjects [@Murphy2025; @Pichardo2023a; @Sammartino2020]. However, empirical methods are usually chosen in small rodents experiments due to the space constraints associated with these models.
+With the growing interest in transcranial ultrasound therapeutic approaches, extensive research has been conducted to develop and validate computational models of acoustic wave propagation through the skull. Although a number of tools and formalisms exist [@Aubry2022a; @Murphy2025], [k-Wave](http://www.k-wave.org) [@Treeby2010] has been widely adopted in the field of ultrasound neurostimulation specifically [@Constans2018; @Verhagen2019; @Yaakub2023]. Acoustic simulations in this context are performed using standalone Matlab scripts for the definition of acoustic sources and domains. Acoustic domains are defined either directly based on CT or pseudo-CT scans [@Aubry2003], or by constructing maps from rasterized brain and skull meshes. Registration of the transducer location relative to targeted brain structures is often achieved using frameless optical tracking neuronavigation systems on human or non-human primate subjects [@Murphy2025; @Pichardo2023a; @Sammartino2020]. However, empirical methods are usually chosen in small rodents experiments due to the space constraints associated with these models.
 
 <!-- // structure atlases + morpho calib -->
 Targeting of brain structures is achieved using reference atlases registered to MRI scans of subjects when available. However for small animals, these images are typically not acquired in a systematic way. Targeted structures coordinates are thus directly evaluated on reference atlases, based on an anatomical landmark such as the Bregma skull suture on rats and mice [@Kleven2023; @Wang2020]. Morphological variability between subjects can compromise experiments if it is not taken into account, however registration of reference atlases to anatomical measurements can be tedious and is rarely reported in rodent studies.
@@ -62,12 +62,12 @@ As of the writing of this article, `CoperniFUS 0.1.0` currently enables:
 1. Facilitated stereotaxic targeting of specific brain structures for a large range of species thanks to the integration of the `BrainGlobe API` [@Claudi2020a].
[…]
