### 2025-06-05 a08902b
readme installation instruction update
