+++ README.md
@@ -41,9 +41,11 @@ For detailed instructions and troubleshooting steps [checkout the documentation 
 1. Install `coperniFUS` using `pip` (if you are using windows, you might need to install Git)
     - `pip install git+https://github.com/Tomaubier/CoperniFUS.git`
 
-        > 1a. *For Linux users:* Install ``libxcb-cursor-dev`` to satisfy ``PyQt6`` requirements ``sudo apt-get install -y libxcb-cursor-dev`` `see <https://stackoverflow.com/questions/77725761/from-6-5-0-xcb-cursor0-or-libxcb-cursor0-is-needed-to-load-the-qt-xcb-platform>`_ for additional details.
+        > 1a. *For Linux users:* Install `libxcb-cursor-dev` to satisfy `PyQt6` requirements `sudo apt-get install -y libxcb-cursor-dev` [see](https://stackoverflow.com/questions/77725761/from-6-5-0-xcb-cursor0-or-libxcb-cursor0-is-needed-to-load-the-qt-xcb-platform) for additional details.
 
-2. Launch `coperniFUS` by running `coperniFUS` in a terminal. Checkout [this page](https://copernifus.readthedocs.io/en/latest/contents/usage.html) for detailed usage instructions.
+        > 1b. *For macOS users:* `k-wave-python` currently requiers `fftw hdf5 zlib libomp` to be installed. This requirement con be satisfied by running `brew install fftw hdf5 zlib libomp` using [Homebrew](https://docs.brew.sh/Installation). [See this k-wave-python issue](https://github.com/waltsims/k-wave-python/issues/549) for additional details.
+
+2. Launch CoperniFUS by running `coperniFUS` in a terminal. Checkout [this page](https://copernifus.readthedocs.io/en/latest/contents/usage.html) for detailed usage instructions.
 
 ## Contributing to the project
 With the aim of making CoperniFUS suitable for a wide range of applications, your feedback and contributions are always welcome! Please refer to the [Contribution Guidelines](CONTRIBUTING.md) if you wish to do so.
