### 2025-06-06 5953475
Configuration switching and creation handler
