+++ coperniFUS/viewer.py
@@ -53,7 +53,7 @@ class Window(pyqtw.QMainWindow):
 
         self.cache = CachedDataHandler('coperniFUSCache')
 
-        self.setWindowTitle(f"CoperniFUS {version('coperniFUS')}")
+        self._update_window_title()
         self.setGeometry(*self.cache.get_attr('viewer.geometry', default_value=[100, 100, 1500, 1000]))
 
         self._init_gui()
@@ -61,7 +61,7 @@ class Window(pyqtw.QMainWindow):
         self._init_modules()
         self.show()
         self.showMaximized()
-    
+
     # ==== Public attributes ====
     
     def get_module_object_from_name(self, module_name=None):
@@ -140,45 +140,53 @@ class Window(pyqtw.QMainWindow):
 
     # --- Cache handler ---
 
-    def switch_cached_settings_file(self, cached_settings_fname):
-        """ Switch between cached configuration files available in cache directory. The list of avalable configuration files can be assessed by calling cached_settings_files. """
-        print('Switching to ', cached_settings_fname)
+    def switch_cached_settings_file(self, cached_settings_fname, force_create_new=False):
+        """ Switch between cached configuration files available in cache directory or create new ones. The list of avalable configuration files can be assessed by calling cached_settings_files. Setting force_create_new to True will force the creation a a new config file by adding a suffix to cached_settings_fname if it already exists."""
+        print(f'> Info: Switching to {cached_settings_fname}')
 
-        if self.cache.is_cached_filename_already_defined(cached_settings_fname):
+        def get_safe_config_name(name):
+            safe_name = copy.deepcopy(name)
+            # Increment suffix until an unused name is found
+            suffix_index = 1
+            while self.cache.is_cached_filename_already_defined(safe_name):
+                safe_name = f'{copy.deepcopy(name)} {suffix_index}'
+                suffix_index += 1
+            return safe_name
 
-            # Freeze wnidow to prevent user interactions during reloading
-            self.setEnabled(False)
-            pyqtw.QApplication.processEvents()
+        if force_create_new and self.cache.is_cached_filename_already_defined(cached_settings_fname):
+            cached_settings_fname = get_safe_config_name(cached_settings_fname)
 
-            # Set new cached file name
-            # self.cache.cached_settings_fname = cached_settings_fname
-            self.cache = CachedDataHandler(cache_dir_name='coperniFUSCache', cached_settings_fname=cached_settings_fname)
+        # Freeze wnidow to prevent user interactions during reloading
+        self.setEnabled(False)
+        pyqtw.QApplication.processEvents()
 
-            # Clear rendered view
-            self.clear_rendered_view()
+        # Set new cached file name
+        # self.cache.cached_settings_fname = cached_settings_fname
+        self.cache = CachedDataHandler(cache_dir_name='coperniFUSCache', cached_settings_fname=cached_settings_fname)
 
-            # Find all dock widgets in the main window
-            docks = self.findChildren(pyqtw.QDockWidget)
-            for dock in docks: # Remove each dock widget
-                self.removeDockWidget(dock)
-                dock.deleteLater()  # Optionally delete the dock widget to free memory
+        # Clear rendered view
+        self.clear_rendered_view()
+        self._update_window_title()
 
-            # Remove all widgets from the status bar
-            status_bar_widgets = self.status_bar.findChildren(pyqtw.QWidget) # Find all QLabel widgets in the status bar
-            for widget in status_bar_widgets:
-                if not isinstance(widget, pyqtw.QSizeGrip): # Bugfix: prevents application crash
[…]
