### 2025-06-06 e8b2be9
Gui cleanup
