### 2025-06-07 a77e8ae
Tutorial section in Docs + minor code cleanup
