+++ coperniFUS/modules/armatures/kwave_armatures.py
@@ -371,7 +371,6 @@ class KWave3dSimulationArmature(STLMeshBooleanArmature):
                 'gl_mesh_edgeWidth': 5
             },
             '_boolean_mask': {
-                '_src_meshes': ['_stl_mesh'],
                 '_boolean_operations': {
                     '1': ['intersection', ['Brain mesh (skull convex Hull)', '_boolean_mask']],
                     '2': ['intersection', ['Skull acoustic window', '_boolean_mask']]
