+++ docs/contents/tutorial.rst
@@ -1,0 +1,609 @@
+Tutorial
+----------
+
+This step by step tutorial demonstrates how to:
+
+- Load and create configurations
+- Model stereotaxic frames
+- Add and register brain atlases based on the location of anatomical landmarks sampled on a subject when no MRI or CT scans are available.
+- Add and register assets such as meshes used for acoustic simulations domain definitions
+- Run k-wave simulations and grab the results for further analysis
+
+.. note::
+    The complete configuration file for this demo is included under ``coperniFUS/examples/assets/Dual arms FUS + rec electrode rat tutorial config.json``.
+
+1. Launch CoperniFUS, load and create configurations
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+Open CoperniFUS by running ``coperniFUS`` in a terminal (or use it in interactive mode as described in `Usage <usage.rst>`_). If no configuration file exist, an empty viewer sill open.
+
+.. image:: /_static/tutorial_assets/tuto_0.png
+
+Create or open configuration files from the *File* menu
+
+.. image:: /_static/tutorial_assets/tuto_1.png
+    :align: center
+    :width: 400px
+
+For the sake of this tutorial select *New Configuration* and input a name.
+
+.. image:: /_static/tutorial_assets/tuto_2.png
+    :align: center
+    :width: 500px
+
+2. Add brain atlases
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+Add a brain atlas by selecting one in the *Brain Atlas* module.
+
+.. image:: /_static/tutorial_assets/tuto_3.png
+
+.. note::
+    To balance rendering performances and accuracy, an atlas subsampling factor can be adjusted.
+
+.. image:: /_static/tutorial_assets/tuto_4.png
+
+Brain structures layers can be added to the viewer.
+
+.. image:: /_static/tutorial_assets/tuto_5.png
+
+Double clicking on the list of loaded layer allows the edition of their color and opacity. The *Tooltip* location can also be set to the centroid of structures assess their location.
+
+.. image:: /_static/tutorial_assets/tuto_6.png
+
+3. Anatomical registration of atlases and assets
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+In the context of studies where access to MRI or CT scans is not possible, CoperniFUS enable the manual registration of atlases and assets such as skull meshes that can be used for acoustic simulations using `k-Wave Python <https://github.com/waltsims/k-wave-python>`_. Based the registration of *anatomical landmarks*, atlases and assets can be automatically transformed to better fit the anatomy of subjects.
+
+In this example, the *Bregma* landmark will be used ad the origin of the coordinate system. A reference image from the literature `[Papp2014] <https://doi.org/10.1016/j.neuroimage.2014.04.001>`_ is used register the Altas and skull meshes relative to its location. Using the Reference Image as Plane module, the image (available in the example/assets folder) can be loaded in the viewer. Click on *Y* to set the plane normal and adjust the pixel size and origin in pixel coordinates.
+
+.. image:: /_static/tutorial_assets/tuto_7.png
+
+To simplify the initial atlas registration procedure, its color can be altered using the aforementioned editor.
+
+.. image:: /_static/tutorial_assets/tuto_8.png 
+
[…]
