### 2025-06-07 69ecadd
Create new config when running tests
