### 2025-06-07 b715975
Tutorial config generator scirpt
