+++ tests/tutorial_config_generator_script.py
@@ -1,0 +1,1036 @@
+
+# %%
+
+""" Script designed to setup CoperniFUS in the tutorial configuration in a programatic way.
+    See 
+"""
+
+steretaxic_frame_buider_dict = {}
+
+# %% --- Bregma to Stereotaxic Frame Origin (Armature) ---
+
+armature_name = 'Bregma to Stereotaxic Frame Origin'
+steretaxic_frame_buider_dict[armature_name] = {}
+steretaxic_frame_buider_dict[armature_name]['armature_type'] = 'Armature'
+
+steretaxic_frame_buider_dict[armature_name]['constants_dict'] = {}
+steretaxic_frame_buider_dict[armature_name]['configuration_dict'] = {
+    '_armature_joints': {
+        'AP Bregma': {
+            'translation_0': {
+                'args': ['x', 0.0784],
+                '_is_editable': True,
+                '_edit_increment': 0.0005,
+                '_color': 'x_RED',
+                '_unit': 'm'
+            }
+        },
+        'ML Bregma': {
+            'translation_0': {
+                'args': ['y', 0.0948],
+                '_is_editable': True,
+                '_edit_increment': 0.0005,
+                '_color': 'y_GREEN',
+                '_unit': 'm'
+            }
+        },
+        'DV Bregma': {
+            'translation_0': {
+                'args': ['z', 0.013],
+                '_is_editable': True,
+                '_edit_increment': 0.0005,
+                '_color': 'z_BLUE',
+                '_unit': 'm'
+            }
+        }
+    }
+}
+
+# --- Skull Mesh (STLMeshArmature) ---
+
+armature_name = 'Skull Mesh'
+steretaxic_frame_buider_dict[armature_name] = {}
+steretaxic_frame_buider_dict[armature_name]['armature_type'] = 'STLMeshArmature'
+
+steretaxic_frame_buider_dict[armature_name]['constants_dict'] = {}
+steretaxic_frame_buider_dict[armature_name]['configuration_dict'] = {
+    '_stl_mesh': {
+        'file_path': 'Pohl2013_coarse.stl',
+        'transform_str': 'Rz90deg Tz-0.292m Tx0.415m Ry4.2deg S.09',
+        'ignore_plane_slicing': False,
+        'ignore_anatomical_landmarks_calibration': False,
+        'gl_mesh_shader': 'boneShader',
+        'gl_mesh_drawEdges': False,
+        'gl_mesh_drawFaces': True,
+        'gl_mesh_edgeColor': [0.9, 0.9, 0.9, 0.7],
+        'gl_mesh_glOptions': 'opaque',
+        'gl_mesh_smooth': False,
+        'gl_mesh_edgeWidth': 2
+    }
+}
+
+# --- Brain Mesh (STLMeshConvexHull) ---
+
+armature_name = 'Brain Mesh'
+steretaxic_frame_buider_dict[armature_name] = {}
+steretaxic_frame_buider_dict[armature_name]['armature_type'] = 'STLMeshConvexHull'
+
+steretaxic_frame_buider_dict[armature_name]['constants_dict'] = {}
+steretaxic_frame_buider_dict[armature_name]['configuration_dict'] = {
+    '_stl_mesh': {
+        'file_path': 'None',
+        'transform_str': 'S1 Rx0deg Tz0mm',
+        'ignore_plane_slicing': False,
+        'ignore_anatomical_landmarks_calibration': True,
+        'gl_mesh_shader': None,
+        'gl_mesh_drawEdges': True,
+        'gl_mesh_drawFaces': False,
+        'gl_mesh_edgeColor': [0.9, 0.9, 0.9, 1],
[…]
