+++ docs/contents/tutorial.rst
@@ -98,7 +98,7 @@ In this example, the following mesh transformation was found: ``Rz90deg Tz-0.292
 4. Add and configure stereotaxic frame Armatures
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-To be used in subsequent operations (mesh boolean, acoustic simulation domain definition, etc.) this mesh needs to be loaded in the *Stereotactic Frame* module. To do that click on *Add armature* and setup an armature of type *STLMeshArmature*.
+To be used in subsequent operations (mesh boolean, acoustic simulation domain definition, etc.) this mesh needs to be loaded in the *Stereotaxic Frame* module. To do that click on *Add armature* and setup an armature of type *STLMeshArmature*.
 
 .. image:: /_static/tutorial_assets/tuto_15.png
     :align: center
@@ -108,7 +108,7 @@ Armatures of any types can be configured using the dedicated editor containing a
 
 .. image:: /_static/tutorial_assets/tuto_16.png 
 
-To load the mesh, parameters under the ``_stl_mesh`` section need to be edited. The ``_armature_joints`` section on the other hand can be discarded in this case. Its purpose will be detailed when defining actual stereotactic frame elements.
+To load the mesh, parameters under the ``_stl_mesh`` section need to be edited. The ``_armature_joints`` section on the other hand can be discarded in this case. Its purpose will be detailed when defining actual stereotaxic frame elements.
 Based on the affine transformation string determined earlier the following configuration can be set.
 
 .. note::
@@ -132,17 +132,17 @@ Based on the affine transformation string determined earlier the following confi
         }
     }
 
-At this point the mesh loaded in the *STL Handler* module can be removed. Note that armature visibility and *Tooltip* location inheritance can be set in the *Stereotactic Frame* module in the same fashion as in the *Brain Atlas* and *Anatomical Landmarks Calibration* modules.
+At this point the mesh loaded in the *STL Handler* module can be removed. Note that armature visibility and *Tooltip* location inheritance can be set in the *Stereotaxic Frame* module in the same fashion as in the *Brain Atlas* and *Anatomical Landmarks Calibration* modules.
 
 .. image:: /_static/tutorial_assets/tuto_17.png 
 
-For demonstration purposes we will now add *Stereotactic Armatures* that would be used for the planing of an experiment where focused ultrasound (FUS) is applied to the *VTA* while monitoring the activity of the *NAc* region using a recording electrode.
+For demonstration purposes we will now add *Stereotaxic Armatures* that would be used for the planing of an experiment where focused ultrasound (FUS) is applied to the *VTA* while monitoring the activity of the *NAc* region using a recording electrode.
 
-To reference anatomical coordinates on the animal during the experiment, we will first define a dummy probe with a needle tip. As mentioned in :ref:`Armature definition procedure <armature_definition_procedure>`, careful measurement of all stereotactic elements need to be performed to obtain good planning and simulation results. An armature definition validation procedure is also described in the :ref:`Armature definition procedure <armature_definition_procedure>` section.
+To reference anatomical coordinates on the animal during the experiment, we will first define a dummy probe with a needle tip. As mentioned in :ref:`Armature definition procedure <armature_definition_procedure>`, careful measurement of all stereotaxic elements needs to be performed to obtain accurate planning and simulation results. For reference, an armature definition validation procedure is also described in the :ref:`Armature definition procedure <armature_definition_procedure>` section.
 
 .. image:: /_static/tutorial_assets/tuto_18.png 
 
[…]
