+++ joss_paper/paper copy.md
@@ -1,83 +1,0 @@
----
-title: 'CoperniFUS: A flexible Python-based GUI for stereotaxic Focused UltraSound (FUS) experiment planning'
-tags:
-  - Python
-  - Experimental Neuroscience
-  - Ultrasound neurostimulation
-authors:
-  - name: Tom Aubier
-    orcid: 0009-0004-5558-8435
-    affiliation: 1 # (Multiple affiliations must be quoted)
-    corresponding: true # (This is how to denote the corresponding author)
-  - name: Sandrine Parrot
-    affiliation: 2
-  - name: Ivan M. Suarez-Castellanos
-    affiliation: 1
-  - name: W. Apoutou N'Djin
-    affiliation: 1
-affiliations:
- - name: LabTAU, INSERM, Centre Léon Bérard, Université Claude Bernard Lyon 1, F-69003, Lyon, France
-   index: 1
- - name: Université Claude Bernard Lyon 1, INSERM, Centre de Recherche en Neurosciences de Lyon CRNL U1028 UMR5292, F-69500 Bron, France
-   index: 2
-date: 7 February 2025
-bibliography: references.bib
-
-# Optional fields if submitting to a AAS journal too, see this blog post:
-# https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
-# aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
-# aas-journal: Astrophysical Journal <- The name of the AAS journal.
----
-
-# Summary
-
-Focused Ultrasound (FUS) is gaining increasing interest for its potential as a minimally invasive and precisely targeted alternative to conventional neurostimulation techniques.
-Although reversible changes in the activity of neural structures have been reported as far back as 1929 [@Harvey1929], comprehensive descriptions of the short- and long-term effects of ultrasound on neural structures are still lacking to achieve neurostimulation with satisfactory levels of control and safety. Delivering well-characterized FUS pulses with a high degree of spatial selectivity, along with local assessment of the state and activity of specific brain regions, is crucial in pursuing this research.
-Unlike most electrophysiology procedures involving compact needle-like probes that can be achieved using a stereotaxic frame limited to three degrees of freedom, FUS experiments on small _in vivo_ models often require the implementation of complex probe layouts to assess the activity of the stimulated neural structures. Treatment planning, evaluation of acoustic parameters through simulations, and post-processing of results often rely on distinct software programs with their own coordinate systems, which greatly complicates the integration, analysis, and interpretation of key information—thus limiting the ability to properly describe the spatiotemporal aspects of FUS neurostimulation dynamics.
-
-`CoperniFUS` is designed to address this challenge by offering a versatile software platform for planning FUS procedures with stereotaxic frames, enabled by a unified coordinate system architecture. Moreover `CoperniFUS` provides the possibility to take anatomical variability into consideration by including the registration of anatomical landmark.
-
-# Statement of need
-
-<!-- // Relevance of studies on FUS-induces microenvironment alteration -->
-In an effort to assess the therapeutic potential of ultrasound neurostimulation, studies have sought  to characterize the effect of ultrasound on the biochemical microenvironment of brain structures after stimulation. While tools like Magnetic Resonance Spectroscopy (MRS) offer a non-invasive way of assessing the concentrations of metabolites _in vivo_, they come with significant limitations. Concentration measurements are non-specific, only representative of the total metabolic, extra- and intracellular quantity of a compound [@Dyke2017]. The spatial selectivity of the method is also limited to minimum voxel volumes of several $cm^3$. Finally, the difficulties arising from the integration of FUS transducers in MRIs are preventing the further investigation of _online_ effects [@Yaakub2023].
-
-<!-- // FUS targetting -->
[…]
