+++ joss_paper/paper.md
@@ -35,7 +35,7 @@ Focused Ultrasound (FUS) is gaining increasing interest for its potential as a m
 Although reversible changes in the activity of neural structures have been reported as far back as 1929 [@Harvey1929], comprehensive descriptions of the short- and long-term effects of ultrasound on neural structures are still lacking to achieve neurostimulation with satisfactory levels of control and safety. Delivering well-characterized FUS pulses with a high degree of spatial selectivity, along with local assessment of the state and activity of specific brain regions, is crucial in pursuing this research.
 Unlike most electrophysiology procedures involving compact needle-like probes that can be achieved using a stereotaxic frame limited to three degrees of freedom, FUS experiments on small _in vivo_ models often require the implementation of complex probe layouts to assess the activity of the stimulated neural structures. Treatment planning, evaluation of acoustic parameters through simulations, and post-processing of results often rely on distinct software programs with their own coordinate systems, which greatly complicates the integration, analysis, and interpretation of key information—thus limiting the ability to properly describe the spatiotemporal aspects of FUS neurostimulation dynamics.
 
-`CoperniFUS` is designed to address this challenge by offering a versatile software platform for planning FUS procedures with stereotaxic frames, enabled by a unified coordinate system architecture. Moreover `CoperniFUS` provides the possibility to take anatomical variability into consideration by including the registration of anatomical landmark.
+Designed around a unified coordinate system architecture, `CoperniFUS` is built to address this challenge by offering a versatile software platform for planning stereotaxic FUS procedures. Moreover `CoperniFUS` provides the possibility to take anatomical variability into consideration by including the registration of anatomical landmark.
 
 # Statement of need
 
