+++ docs/index.rst
@@ -13,7 +13,7 @@ In the quest for a better control and understanding of the mechanisms of ultraso
 Designed around a unified coordinate system architecture, `CoperniFUS` is built to address this challenge by offering a versatile software platform for planning stereotaxic FUS procedures.
 
 .. warning::
-   Like any other numerical modeling technique, the quality of the results produced by CoperniFUS depends entirely on the choice of input parameters. This software, developed in the context of my PhD, is also subject to bugs. It is therefore of the responsibility of the user to design and perform careful assessments of the results validity. If you identify any significant issues, please document them `in GitHub's issue section <https://github.com/Tomaubier/CoperniFUS/issues>`_.
+   Like any other numerical modeling technique, the quality of the results produced by CoperniFUS depends entirely on the choice of input parameters. This software, developed in the context of my PhD, is also subject to bugs. It is therefore of the responsibility of the user to design and perform careful assessments of the results validity.  If you identify any issues, please report them `in GitHub's dedicated section <https://github.com/Tomaubier/CoperniFUS/issues>`_.
 
 .. image:: /_static/CoperniFUS_screenshot.png
 
