+++ docs/index.rst
@@ -10,6 +10,7 @@ CoperniFUS documentation
 CoperniFUS: A flexible Python-based GUI for stereotaxic experiment planning.
 
 In the quest for a better control and understanding of the mechanisms of ultrasound neuromodulation and neurostimulation, the planning of experiments, evaluation of acoustic parameters via simulations, and post-processing of results often rely on distinct software programs with their own coordinate systems, which greatly complicates the integration, analysis, and interpretation of key information.
+
 Designed around a unified coordinate system architecture, `CoperniFUS` is built to address this challenge by offering a versatile software platform for planning stereotaxic FUS procedures.
 
 .. warning::
