### 2025-06-07 772bf4d
Minimum requirements section in docs
