### 2025-06-08 2433684
download example assets on pip install
