+++ pyproject.toml
@@ -3,8 +3,8 @@ requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools.packages.find]
-include = ["coperniFUS*"]
-exclude = ["joss_paper*", "docs*", "*examples*"]
+include = ["coperniFUS*", "*examples*"]
+exclude = ["joss_paper*", "docs*"]
 
 [project]
 name = "coperniFUS"
