### 2025-06-08 670fcdf
download example assets on pip install
