### 2025-06-08 c438601
download example assets on pip install debug
