+++ pyproject.toml
@@ -3,8 +3,12 @@ requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools.packages.find]
-include = ["coperniFUS*", "coperniFUS/examples/**"]
+include = ["coperniFUS*"]
 exclude = ["joss_paper*", "docs*"]
+
+[tool.setuptools]
+include-package-data = true
+package-data = { "coperniFUS" = ["examples/*"] }
 
 [project]
 name = "coperniFUS"
