### 2025-06-08 76d6485
inculde example assets in pip download
