### 2025-06-08 333106b
Qthread bug fix
