### 2025-06-08 16153a7
interactive example update
