### 2025-06-09 #8128 @Tomaubier
@Julie-Fabre and @richardkoehler as you may have seen in https://github.com/Tomaubier/CoperniFUS/issues/6 and https://github.com/Tomaubier/CoperniFUS/issues/7 I have tried to address all of your comments. Please let me know if the changes I implemented still need to be improved upon.
Best regards,
Tom
