### 2025-06-10 Tomaubier/CoperniFUS#6 @Tomaubier
Hi,
1. From 8df8ff623443154ad622f1a0c79bd6d4bab2abc7 onwards, the proper default orientation of atlases is ensured thanks to the [brainglobe-space](https://brainglobe.info/documentation/brainglobe-space/index.html) package from BrainGlobe. As described [in the tutorial](https://copernifus.readthedocs.io/en/latest/contents/tutorial.html#anatomical-registration-of-atlases-and-assets) the alignment of atlases and skulls is done manually using by specifying sets of affine transformations expressed in the [transformation from strings syntax](https://copernifus.readthedocs.io/en/latest/contents/usage.html#affine-transformations-from-strings-syntax). At some point I considered the possibility to display the coordinates of the Tooltip in the Altas reference frame. However to my knowledge, brainglobe does not provide relations between the image spaces of its altases and the stereotaxic coordinates. From [what I've read and understood](https://community.brain-map.org/t/how-to-transform-ccf-x-y-z-coordinates-into-stereotactic-coordinates/1858) image space coordinates like CCF are not used for targeting so I scrapped the idea. Let me know if you think that this addition would be useful.

2. Do you have resources to evaluate minimum hardware requirements? I added a very general sentence in the [installation section of the docs](https://copernifus.readthedocs.io/en/latest/contents/installation.html#minimum-requirements) however given the variety of potential uses, it is not very clear to me how to provide specific recommendations.

3. Currents tests ensures that the caching system used in the foundations of the software work as expected. Trimesh and k-wave interfaces (wrappers) are tested by themselves although the k-wave wrapper for axisymmetric simulation is skipped for now as their is [a known bug preventing it from running on Darwin systems](https://github.com/waltsims/k-wave-python/issues/470). The Tooltip, AnatomicalLandmarksCalibration, StereotaxicFrame and BrainAtlas modules are tested by simulating GUI interactions using `pytest-qt`. Finally a pipeline equivalent to what's presented in the tutorial (with dummy meshes and coarse simulation grids to prevent unnecessary loading of Github's servers) is runned through and the results compared to reference values encoded as hashes or loaded from `tests/reference_test_assets`.

Let me know if you have any suggestions!
