### 2025-06-10 Tomaubier/CoperniFUS#6 @Julie-Fabre
Hi Tom, thank you! That tutorial is great! Good work.
- I agree that CCF is not used for targetting (or should not be anyways) - and you don't provide any initial alignment so that is good for me (I think I had misunderstood there!).
- I agree hardware requirements can be hard to evaluate. If you can specify for instance whether a GPU is necessary (I think it is not) that can be nice, but otherwise it is fine as is.
- Tests cover important parts of the code and they work for me.

In summary, that answers my last questions. Great work and I am happy to recommend your package for publication in JOSS :tada:
