### 2025-06-12 59a3342
Bumping trimesh to v4.6.12 -> dict_to_path_patched is depricated
