+++ coperniFUS/modules/armatures/kwave_armatures.py
@@ -175,11 +175,17 @@ mesh = extrusion.to_mesh()
         if self._axisym_domain_mesh is None:
             accessible_globals_names = [
                 'trimesh', 'np',
-                'dict_to_path_patched'
+                'dict_to_path_patched'  # Depricated in v0.1.1
             ]
 
-            accessible_globals = {accessible_glob_name: globals()[accessible_glob_name] for accessible_glob_name in accessible_globals_names}
-            accessible_globals = {**accessible_globals, **bool_mask_params}
+            accessible_globals = {
+                accessible_glob_name: globals()[accessible_glob_name]
+                for accessible_glob_name in accessible_globals_names
+            }
+            accessible_globals = {
+                **accessible_globals, **bool_mask_params,
+                'dict_to_path': trimesh.path.exchange.misc.dict_to_path
+            }
 
             # run trimesh script
             try:
@@ -401,7 +407,7 @@ path_2d_dict = {
 }
 
 path_2d_from_dict = trimesh.path.exchange.load.load_path(
-    dict_to_path_patched(path_2d_dict)
+    dict_to_path(path_2d_dict)
 )
 
 extrusion = path_2d_from_dict.extrude(threeD_domain_z_size)
@@ -775,7 +781,7 @@ path_2d_dict = {
 }
 
 path_2d_from_dict = trimesh.path.exchange.load.load_path(
-    dict_to_path_patched(path_2d_dict)
+    dict_to_path(path_2d_dict)
 )
 
 extrusion = path_2d_from_dict.extrude(threeD_domain_z_size)
@@ -1141,11 +1147,17 @@ mesh = extrusion.to_mesh()
         if self._axisym_domain_mesh is None:
             accessible_globals_names = [
                 'trimesh', 'np',
-                'dict_to_path_patched'
+                'dict_to_path_patched'  # Depricated in v0.1.1
             ]
 
-            accessible_globals = {accessible_glob_name: globals()[accessible_glob_name] for accessible_glob_name in accessible_globals_names}
-            accessible_globals = {**accessible_globals, **bool_mask_params}
+            accessible_globals = {
+                accessible_glob_name: globals()[accessible_glob_name]
+                for accessible_glob_name in accessible_globals_names
+            }
+            accessible_globals = {
+                **accessible_globals, **bool_mask_params,
+                'dict_to_path': trimesh.path.exchange.misc.dict_to_path
+            }
 
             # run trimesh script
             try:
