+++ pyproject.toml
@@ -42,7 +42,7 @@ dependencies = [
     "opencv-python==4.10.0.84",
     "pymeshfix==0.17.1",
     "manifold3d==3.1.1",
-    "trimesh==4.6.10",
+    "trimesh==4.6.12",
     "matplotlib==3.10.0",
     "napari==0.6.1",
     "onnxruntime==1.19.2",
