### 2025-06-12 67e2be9
TrimeshScriptArmature default script typo + default color dark/light mode compatibility
