+++ coperniFUS/modules/interfaces/trimesh_interfaces.py
@@ -9,7 +9,7 @@ class TrimeshHandler:
         'gl_mesh_shader': 'viewNormalColor',
         'gl_mesh_drawEdges': False,
         'gl_mesh_drawFaces': True,
-        'gl_mesh_edgeColor': (.9, .9, .9, 1),
+        'gl_mesh_edgeColor': (.5, .5, .5, .7),
         'gl_mesh_glOptions': 'opaque',
         'gl_mesh_smooth': False,
         'gl_mesh_edgeWidth': 5,
@@ -200,7 +200,7 @@ class StlHandler(TrimeshHandler):
         'gl_mesh_shader': 'viewNormalColor',
         'gl_mesh_drawEdges': False,
         'gl_mesh_drawFaces': True,
-        'gl_mesh_edgeColor': (.9, .9, .9, 1),
+        'gl_mesh_edgeColor': (.5, .5, .5, .7),
         'gl_mesh_glOptions': 'opaque',
         'gl_mesh_smooth': False,
         'gl_mesh_edgeWidth': 5,
@@ -319,7 +319,7 @@ class StlHandler(TrimeshHandler):
         'gl_mesh_shader': 'viewNormalColor',
         'gl_mesh_drawEdges': False,
         'gl_mesh_drawFaces': True,
-        'gl_mesh_edgeColor': (.9, .9, .9, 1),
+        'gl_mesh_edgeColor': (.5, .5, .5, .7),
         'gl_mesh_glOptions': 'opaque',
         'gl_mesh_smooth': False,
         'gl_mesh_edgeWidth': 5,
