+++ docs/conf.py
@@ -30,7 +30,7 @@ shutil.copyfile(source_readme, dest_readme)
 project = 'CoperniFUS'
 copyright = '2025, Tom Aubier'
 author = 'Tom Aubier'
-release = '0.1.0'
+release = '0.1.1'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
