+++ pyproject.toml
@@ -19,7 +19,7 @@ license-files = ["LICENSE"]
 authors = [
     { name = "Tom Aubier", email = "tom.aubier@inserm.fr" },
 ]
-version = "0.1.0"
+version = "0.1.1"
 requires-python = ">=3.12"
 dependencies = [
     "PyQt6==6.7.0",
