+++ coperniFUS/modules/_stereotaxic_frame_helper_classes.py
@@ -1,4 +1,5 @@
 from coperniFUS import *
+import functools
 
 
 class PythonSyntaxHighlighter(pyqtg.QSyntaxHighlighter):
