+++ coperniFUS/modules/armatures/kwave_armatures.py
@@ -3,7 +3,8 @@ from coperniFUS.modules.armatures.mesh_armatures import STLMeshBooleanArmature
 from coperniFUS.modules.armatures.base_armature import Armature
 from coperniFUS.modules.interfaces.kwave_interfaces import *
 from coperniFUS.modules.interfaces.trimesh_interfaces import *
-
+import matplotlib.pyplot as plt
+import threading, scipy
 
 class KwaveAShomogeneousSimulationArmature(Armature):
 
