+++ docs/requirements.txt
@@ -6,6 +6,7 @@ sphinx-autodoc-typehints==3.0.1
 sphinx_mdinclude==0.6.2
 myst-parser==4.0.1
 
+PyOpenGL==3.1.9
 PyQt6==6.7.0
 numpy==1.26.4
 scipy==1.14.1
@@ -14,7 +15,6 @@ pyqtgraph==0.13.7
 brainglobe_atlasapi==2.0.7
 scikit-spatial==7.2.0
 si-prefix==1.3.3
-numpy-stl==3.1.1
 shapely==2.1.1
 rtree==1.3.0
 mapbox_earcut==1.0.1
@@ -28,7 +28,5 @@ pymeshfix==0.17.1
 manifold3d==3.1.1
 trimesh==4.6.12
 matplotlib==3.10.0
-napari==0.6.1
 onnxruntime==1.19.2
 k-Wave-python==0.4.0
-.
