### 2025-06-17 db40b05
docs requirements typo fix
