+++ docs/requirements.txt
@@ -30,3 +30,4 @@ trimesh==4.6.12
 matplotlib==3.10.0
 onnxruntime==1.19.2
 k-Wave-python==0.4.0
+.
