### 2025-06-18 Tomaubier/CoperniFUS#7 @richardkoehler
@Tomaubier
Thanks for the response - everything seems to work fine now. I think your solution with manual downloads is reasonable if the automatic downloads are not reliable enough.
Fyi all tests pass on my machine as well (Win11).
I will give the green light for publication from my side now!
