### 2025-06-18 83d6bb3
Docs/paper figures + ORCIDs updates
