### 2025-06-18 b5bbe45
Proper CoperniFUS version in paper
