+++ joss_paper/paper.md
@@ -60,7 +60,7 @@ In this context we developed `CoperniFUS`, a modular software tool offering a co
 
 ![Overview of `CoperniFUS`'s Graphical User Interface architecture.\label{fig:example}](figures/CoperniFUS_ui_breakdown.png)
 
-As of the writing of this article, `CoperniFUS 0.1.0` currently enables:
+As of the writing of this article, `CoperniFUS 0.1.2` currently enables:
 1. Facilitated stereotaxic targeting of specific brain structures for a large range of species thanks to the integration of the `BrainGlobe API` [@Claudi2020a].
 2. Arbitrary stereotaxic frames can be constructed using a flexible dictionary-based editor. Discrete stereotaxic frame elements (referred as `Armatures`) can be associated in a hierarchical tree structure, allowing operations such as mesh boolean operations or acoustic simulations to be performed in any coordinate reference frame.
 3. In the absence of whole head CT or MRI scans that allow for accurate atlas registration, `CoperniFUS` provides a means of minimizing the targeting errors that can arise from anatomical variability. To that end, 3D objects such as atlases or skull meshes can inherit from an anatomically calibrated coordinate frame whose scale and orientation can be easily matched to subjects using anatomical landmarks such as Lambda and Bregma for rodents.
@@ -69,7 +69,7 @@ As of the writing of this article, `CoperniFUS 0.1.0` currently enables:
 
 # Software architecture
 
-The software is composed of a viewer graphical interface based on `PyQt6`, hosting a `pyqtgraph` 3D viewport. Modules are integrated to the GUI as dockable elements and allow the manipulation of objects that can be visualized in the viewport. At its core, `CoperniFUS 0.1.0` is equipped with three built-in modules: *1.* the `Tooltip module`, which allows the user to input coordinates, visualized in the viewer by an axis triad. Section views of items displayed in the viewport can be performed along planes normal to its $x$, $y$ and $z$ axes. Furthermore, Tooltip coordinates can be referenced by any modules, facilitating anatomical landmark registration, or assessment of distances and locations in any coordinate frames. *2.* The `Stereotaxic frame module` handles the creation of arbitrary stereotaxic frames using `armatures` objects that can be combined in a hierarchical structure. At its core, an `armature` consists of a series of affine transformation operations (translation & rotation) that allow for the modeling of rigid frames. Specialized armature objects also exist to perform operations in the coordinates of stereotaxic frame elements. Thus, `trimesh armatures` have been developed to perform import `.stl` meshes, create arbitrary geometries programmatically or perform operations such as booleans or `convexhull` mesh generation. A dedicated armature object leveraging the [Python implementation of k-Wave](https://github.com/waltsims/k-wave-python) [@Yagubbbayli2024] has been designed to evaluate pressure field in the context of FUS neurostimulation studies. From a programmer perspective, specialized armatures inherit from a common `Armature` class, which implements the coordinate transformation logic and user interface aspects. *3.* Finally, the `Anatomical registration helper` module handles spatial transformations required to match the location, rotation and scale of atlases and objects such as skull meshes or reference images to anatomical landmarks acquired experimentally. As represented on \autoref{hierarchical_coordinate_system}, coordinate frame associated to modules and armatures used throughout the software are part of a common hierarchical system. Any objects displayed in the viewer can thus be easily set to inherit from the anatomically calibrated coordinate system.
[…]
