### 2025-06-18 8c5f1ca
v0.1.2 version bump + co-corresponding author
