+++ pyproject.toml
@@ -19,7 +19,7 @@ license-files = ["LICENSE"]
 authors = [
     { name = "Tom Aubier", email = "tom.aubier@inserm.fr" },
 ]
-version = "0.1.1"
+version = "0.1.2"
 requires-python = ">=3.12"
 dependencies = [
     "PyOpenGL==3.1.9",
