### 2025-06-18 #8128 @Tomaubier
Dear @Kevin-Mattheus-Moerman,
On the basis of a discussion with Sandrine Parrot, second author on this article, she informed us that she would prefer to be mentioned in the acknowledgments section. She argued that, as a biochemist and expert in microdialysis, her contribution in the present study did not give her sufficient legitimacy to be listed as one of the authors of CoperniFUS and wishes to be given co-authorship in future articles employing the tool in the context of neurochemical experiments.

To be fully transparent, I would like to clarify the role of each author currently involved:
- @Tomaubier: conceptualisation, software development, writting
- Ivan M. Suarez-Castellanos: conceptualisation, funding acquisition, proof-reading, supervision
- W. Apoutou N'Djin: funding acquisition, proof-reading, supervision
- Sandrine Parrot: technical support, microdialysis training

In the hope that this change does not compromise the approval process at this stage, I would ask you to consider this request.
Best regards,
Tom Aubier
