### 2025-07-11 #8128 @richardkoehler
> [@richardkoehler](https://github.com/richardkoehler) You recommended acceptance but left the scholarly effort box unticked. Can you tick this now or explain/comment on this? Thanks.

Actually, that must have been a misclick, I can't explain it otherwise. It would have been one of the first boxes that I ticked off. I went through all boxes again and made sure they are all ticked
