### 2025-07-11 #8128 @Tomaubier
> [@Tomaubier](https://github.com/Tomaubier) can you please work on the below:
>
> * [ ]  Edit the archive title so it matches the paper title.[ ]  Edit the archive listed license to include the `v` as well so it will read `version v0.1.2` on Zenodo. Thanks.

It should be done now, let me know if I made any mistakes.
