### 2025-07-11 #8128 @Tomaubier
> [@Tomaubier](https://github.com/Tomaubier) my apologies your version should actually not have the `v` (it is not there on the GitHub release), so it should be edited back to be `0.1.2` on the archive. I'll fix it here now too.

No worries, it is now back to 0.1.2
