### 2025-07-11 #8128 @Tomaubier
> ⚠️ [Error preparing paper acceptance](https://github.com/openjournals/joss-papers/actions/runs/16221167219?check_suite_focus=true). The generated XML metadata file is invalid.
>
> ID ref-Yaakub2023 already defined
> ID ref-Min2011 already defined

@Kevin-Mattheus-Moerman do I need to change anything in the paper source to fix this error?
