### 2025-07-11 #8128 @xuanxu
`Yaakub2023` and `Min2011` entries appear twice in the [bib file](https://github.com/Tomaubier/CoperniFUS/blob/main/joss_paper/references.bib).
