### 2025-07-11 b7fc258
bib file cleanup
