+++ joss_paper/references.bib
@@ -231,23 +231,6 @@
   file = {/Users/tomaubier/Library/Mobile Documents/iCloud~md~obsidian/Documents/UsNeuStim/Files/Verhagen2019.pdf}
 }
 
-@article{Yaakub2023,
-  title = {Transcranial Focused Ultrasound-Mediated Neurochemical and Functional Connectivity Changes in Deep Cortical Regions in Humans},
-  author = {Yaakub, Siti N. and White, Tristan A. and Roberts, Jamie and Martin, Eleanor and Verhagen, Lennart and Stagg, Charlotte J. and Hall, Stephen and Fouragnan, Elsa F.},
-  year = {2023},
-  month = sep,
-  journal = {Nature Communications},
-  volume = {14},
-  number = {1},
-  pages = {5318},
-  issn = {2041-1723},
-  doi = {10.1038/s41467-023-40998-0},
-  urldate = {2023-10-18},
-  abstract = {Abstract             Low-intensity transcranial ultrasound stimulation (TUS) is an emerging non-invasive technique for focally modulating human brain function. The mechanisms and neurochemical substrates underlying TUS neuromodulation in humans and how these relate to excitation and inhibition are still poorly understood. In 24 healthy controls, we separately stimulated two deep cortical regions and investigated the effects of theta-burst TUS, a protocol shown to increase corticospinal excitability, on the inhibitory neurotransmitter gamma-aminobutyric acid (GABA) and functional connectivity. We show that theta-burst TUS in humans selectively reduces GABA levels in the posterior cingulate, but not the dorsal anterior cingulate cortex. Functional connectivity increased following TUS in both regions. Our findings suggest that TUS changes overall excitability by reducing GABAergic inhibition and that changes in TUS-mediated neuroplasticity last at least 50\,mins after stimulation. The difference in TUS effects on the posterior and anterior cingulate could suggest state- or location-dependency of the TUS effect---both mechanisms increasingly recognized to influence the brain's response to neuromodulation.},
-  langid = {english},
-  file = {/Users/tomaubier/Library/Mobile Documents/iCloud~md~obsidian/Documents/UsNeuStim/Files/Yaakub2023.pdf}
-}
-
 @article{Aubry2003,
   title = {Experimental Demonstration of Noninvasive Transskull Adaptive Focusing Based on Prior Computed Tomography Scans},
   author = {Aubry, J.-F. and Tanter, M. and Pernot, M. and Thomas, J.-L. and Fink, M.},
@@ -304,23 +287,6 @@
   langid = {english},
   pmid = {32386544},
   file = {/Users/tomaubier/Library/Mobile Documents/iCloud~md~obsidian/Documents/UsNeuStim/Files/Wang2020.pdf}
-}
-
-@article{Min2011,
-  title = {Focused Ultrasound Modulates the Level of Cortical Neurotransmitters: {{Potential}} as a New Functional Brain Mapping Technique},
-  shorttitle = {Focused Ultrasound Modulates the Level of Cortical Neurotransmitters},
-  author = {Min, Byoung-Kyong and Yang, Po Song and Bohlke, Mark and Park, Shinsuk and R.Vago, David and Maher, Timothy J. and Yoo, Seung-Schik},
-  year = {2011},
-  month = jun,
-  journal = {International Journal of Imaging Systems and Technology},
-  volume = {21},
-  number = {2},
-  pages = {232--240},
-  issn = {08999457},
-  doi = {10.1002/ima.20284},
-  urldate = {2022-09-19},
-  langid = {english},
-  file = {/Users/tomaubier/Library/Mobile Documents/iCloud~md~obsidian/Documents/UsNeuStim/Files/Min2011.pdf}
 }
 
 @misc{Olaitan2024,
