### 2025-07-24 #8128 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
